fix bug 22086, by storing map data correctly in savegame summaries
The bug was that minimap previews for save games were not appearing,
or that the wrong ones were appearing.

It seems that because of more relics from this commit / related,

https://github.com/wesnoth/wesnoth/commit/c17d84b12

the savegame summaries generated by "extract savegame summary" were
bugged and didn't store the map data correctly. This bug wasn't
obvious because the save game load dialog tries to scavenge what
ever the current game config is for map data if it can't find this,
but all of the saveindex data was being generated incorrectly.

After this commit all the save files seem to work correctly,
except for start-of-scenario saves for reasons that aren't clear
to me, and which I think is a separate issue. (In that case,
something is going wrong with the game_config lookup process most
likely.)

This only appears to fix the bug for normal saves, autosaves, and
replay saves.


Former-commit-id: 61531323a8bb2d34118509ee72fa53f8c52515c1
Former-commit-id: f1281f7f3209d92609017633831fb4763e339ebc
Former-commit-id: 37c52f7a1dc7d6b9f392cc304c901014ca3a7cf9
Former-commit-id: efbb2ae9f284cdf434858ddaa6316c4d9d4541bd [formerly 6fdea68f30456f90023c5b2c0e777eaa737f9387]
Former-commit-id: 4663e6a575eb7de3a20a61ae6136f1ef8a0c1adc