check lobby_sounds preference for the GUI 1 lobby also
This fixes a bug reported on forums here:
http://forums.wesnoth.org/viewtopic.php?f=4&t=40660#p572131

The lobby sounds pref now affects the GUI 1 and 2 mp lobbies,
but it only affects the lobby screen, not mp::wait or mp::connect
etc.

This check is implemented by trying to dynamic cast 'this' mp::ui
object to a mp::lobby, at the time when we would play the arrive
or leave sounds.

The lobby sounds pref does *not* block the sounds due to server
messages, messages from a friend, or whispers.


Former-commit-id: 52189b641922e03742a2170e8b17a263d9d53f18
Former-commit-id: 11fb66aedf399aa2b4366f158213d3b6ad85c26c
Former-commit-id: 24c1ee581fe96a90e2611351ce086577833b6682
Former-commit-id: d33a70ca9f262a3e35fe0eee05fe09f949ce45f4 [formerly 69a919f3b9da9c76a1cec7f5f64b7262262c0f24]
Former-commit-id: 9f4614d20e8e85fef65e0b4bbb9c03b0ce25789b