check lobby_sounds preference for the GUI 1 lobby also
This fixes a bug reported on forums here:
http://forums.wesnoth.org/viewtopic.php?f=4&t=40660#p572131

The lobby sounds pref now affects the GUI 1 and 2 mp lobbies,
but it only affects the lobby screen, not mp::wait or mp::connect
etc.

This check is implemented by trying to dynamic cast 'this' mp::ui
object to a mp::lobby, at the time when we would play the arrive
or leave sounds.

The lobby sounds pref does *not* block the sounds due to server
messages, messages from a friend, or whispers.


Former-commit-id: 316fc295bc95facee935fa517a11ec248c1216f8
Former-commit-id: 84c7fca7d026ad2060c4facc6d7cd206b4cb3ccf
Former-commit-id: d3ac883a58af5282e6b93806aec1d87d2a0d0afe
Former-commit-id: b016ebc543171686a2b3c1c47fc7bb4baebea94e [formerly f613c357ab09e8f4fe275df6797d9f366122af04]
Former-commit-id: 06bdad354265e0066e5c6264a54478def4701f90