give a warning when notifications enabled but not available
The "disable notifications" option in advanced preferences allows
the user to toggle desktop notifications on and off. However, if
the game was not compiled with support for these, the option won't
do anything.

In this revision, if disable_notifications is set to false, but
notifications are not available, a dialog is launched which
explains that this executable was not compiled with support for
notifications.

The preference *is* toggled however -- this is because all versions
of wesnoth use the same preference file (at least under linux), so
we would rather to allow the user to change the file at will --
a different wesnoth binary which they have may indeed support
notifications. So we only give a warning but allow them to toggle.


Former-commit-id: 877dbacbe0d764e6ff2bba440e79429845e99208
Former-commit-id: fbd76203a7f855143a0299accaf78e0c33f83204
Former-commit-id: 861269b394a784aa7722d5e02b52c123b886160b
Former-commit-id: 74032153e6649c9f529f37beabecb782f776ada2 [formerly 014370b3bd590d5ac0e02564a7a5fcfb03f254f4]
Former-commit-id: dbb2bf8faccf2f5182ac84f72b311a74da8fb40d