add "Defense Capped" col. to terrain table in help page for units
If a unit has a "negative" modifier on one of its terrain defenses,
for the basic terrain defense types that have been discovered and
will be displayed on the current help page, a column will be
generated "Defense Capped" which indicates which terrain types
have the negative modifier. In each row the column contains either
'-' or 'yes'.


Former-commit-id: e7a62bd8f095b7c9cef02e0138ec815c5da4d277
Former-commit-id: 0a896cee3432c2ac148aded786c2205e841a7d6f
Former-commit-id: 3eafe54d5f1c3cb08ec9a93c3a8e17e633097ea7
Former-commit-id: 9299ace29a8f94a11eed633bdf0a2ccb25b73d56 [formerly f3257c43ff4ab63aafaeb6fc3d3cfc72a7efe9c9]
Former-commit-id: ef1507a5906ef9a0a47504b09d5566897a8067e2