make the help smarter -- musthave traits without filters are applied
When a unit type has "musthave" traits which always apply, by
virtue of not having a filter, and they affect movement type,
resistance, or jamming / vision data, the "base" unit type data
is updated with these effects when displayed in the help.

The only mainline example I know of where this makes a change is
that the "feral" trait results in a "defense capped" column being
generated by the help. But it can be tested that if core files
are modified so that a musthave trait which always applies makes
a change to the movement type or resistances, then this will be
displayed in the help page for that unit type.

It might also be handy for certain UMC which makes use of racial
traits to modify stats.

Note that we don't at the moment:
- Update health or mp of the unit, or it's experience, according
to such trait effects.
- Update the unit with added / removed abilities due to such trait
effects.

So the help could be smarter still.


Former-commit-id: 0396588dcaeff46156c33fa82436afdc5ed1baca
Former-commit-id: 90ad1dea52a18eba22320e16715aa5332656a49d
Former-commit-id: 1633074fcbec9180a157d20c25834e77c61e5a72
Former-commit-id: d5f44541614524e3dbd9891fe319fac85be3b2fb [formerly 4a0942d995b57b31a47451c5fdb4ee47c1f7a676]
Former-commit-id: c9ef8940cf7b360269acae9c52c93b1094295d45