add description of "Best/worst defense/movement" to terrain help
When viewing a terrain description (ctrl-t), or looking at the
description of a mixed terrain in the help, wesnoth will parse
the defense and movement alias strings, and format it in the
translated language, color coding the words "Best of " in green
and "Worst of " in red. These descriptions become parenthesized
with subexpressions if there is a complicated terrain with many
types and multiple alternations.


Former-commit-id: c7a06b4644de44d55017fa667ab682acbf332c1b
Former-commit-id: ffaf10e4389c9ca425519548a95b63dea4d734f8
Former-commit-id: 0479eee24b00a1a59ccf7f30416f55eb92c0f6ec
Former-commit-id: 48b32b49470c80a5fa742d3d3b8b271491be82f3 [formerly 39685ac11bd0b1655b2677c8c18e44b7bcd7b330]
Former-commit-id: 050843d94866c988d43bea54b64dfc4fc67b3720