experimental implementation of sp advanced settings dialog
Era and mod selection is functional, as well as options to enable mp_configure and mp_create dialogs.
There is still an issue of realoading config. If you load multiplayer, then exit, then load campaign, there is no config reload, so the multiplayer eras/mods stay in config and are selectable. However, since state.classification().campaign_type is not set to multiplayer, when config is reloaded for a specific campaign, multiplayer won't be defined and therefore the selected era may no longer be available. So, we need some way to handle this case better. Also, it may be desirable to add a button to advanced settings which would change the campaign_type to multiplayer and reload (if we actually did want to load some mp era). Another potential option for this dialog could be IO type (to make the game networked, etc).


Former-commit-id: b493b3216f3c3ec421931e9bebf5923ea1461bbc
Former-commit-id: cacda8a4389c92397f58b1477b2e5d87d55832d0
Former-commit-id: b692bed7b6e8d34067d2ca4fe005e384da0bd199
Former-commit-id: a3e7a44233525551cd9e2ae23868b921f36185ac [formerly db93d0f90efc3f563c5556e1e793049eb9b1f4a8]
Former-commit-id: f8436580b155b08a213c857857e244b032e373e6