partly revert e454252f2
there are multiple reasons why i think this should be reverted:
1) this wasn't tested enough, especialy we don't even know, whether all mainline campaigns are compatible
(this can be easily fixed, but only if one know which campaigns are effected which requires testing)
2) after renaming remove_from_carryover_on_leaders_loss to remove_from_carryover_on_defeat it sounds wrong that this can result in removing units on victory.
3) this seemed to break many UMC addons.

Note that this only removes one check for defeat_condition before [endlevel].
Especialy if for example side 1 gets defeated by defeat_condition in turn 1 but the game still goes on due to multiple other non defeated sides, but then the game ends with [endlevel] result=victory in turn 2, the units will be removed from carryover (unless specified otherwise)
This was also the pre 1.11.13 behaviour.
Also note that we cannot remove the 'defeated' side in case of [endlevel]result=victory, becasue [endlevel] doesn't sepcify which sides have victory/defeat.


Former-commit-id: 7e178d21543e0ed1fb175d11fe5c22f840d1ae70
Former-commit-id: 932f3510dbc7d0061feb2d3f4e140d22664d63ac
Former-commit-id: c1a6c8f081c5ac5c485da0ec07df1c77ba909c90
Former-commit-id: 1ff4e7cd827cd04b67b7ea193bd7c4dc926cde66 [formerly 038b11c8ad23736c85947625d47da2be2eb77b7b]
Former-commit-id: ea7e63bf34871696f592b0eecbfe7d2cdccc370b