fix copy assignment operator for unit class to work with refcount
This issue was spotted by gfgtdf. It doesn't appear that we
actually use this function -- prior to this commit, I tested a
version which had "assert(false)" at the start of this function,
and it was never triggered in normal play. But it's possible that
some lua command or obscure WML tag does use it.

It might also be a good idea to make the unit class noncopyable.


Former-commit-id: b17facec1e413e155a150fe3da1212ad75e925ad
Former-commit-id: 54a18b61d2b14eb1daed17a5012601d71346b352
Former-commit-id: 167dbc40a37d135f8df714539c09cd6d45ec1f30
Former-commit-id: baad929d6440a1c3d1931ceb1947366743725da1 [formerly 2659f40666af8fc0b479e51e207515b5b6f7bfd9]
Former-commit-id: 5e5ce8888bf610a8e8a0d618773d3097ecf51a72