geography.cfg: correct wmllint spelling for Castelfranc and Rumyr
These towns were misspelled in their encyclopedia entries, compared to
the Wesnoth map. All of the instances in my UMC collection follow the map
spellings, so I am not leaving the incorrect spelling as an alternative.

Also, there were spelling rules for both "Memog" and Melmog on separate
lines. No reason to keep the mistake around.


Former-commit-id: fb02f1a4b147b60979cdd7bb52cc3af15038739b
Former-commit-id: 94cff864463cec590f0fbea58ee57aa51897882b
Former-commit-id: 10e791baef072096fb7772a76109d279813f17f2
Former-commit-id: 7e7bfb445f476665dee0a0367fd53741955d5af0 [formerly b9e6aea64b58148ea0a931584dc8db737b5593cb]
Former-commit-id: 3a416116b4571006b9a4340a29ae959d3ef98578