geography.cfg: correct wmllint spelling for Castelfranc and Rumyr
These towns were misspelled in their encyclopedia entries, compared to
the Wesnoth map. All of the instances in my UMC collection follow the map
spellings, so I am not leaving the incorrect spelling as an alternative.

Also, there were spelling rules for both "Memog" and Melmog on separate
lines. No reason to keep the mistake around.


Former-commit-id: d3a57767dec27e0480969e62ba21a9b898c53cfc
Former-commit-id: 5f2e04aecf174132524d2a3be03fa03eaa94214f
Former-commit-id: 99204709138ff3ed062eb0d786c03d31e758a8e9
Former-commit-id: 50228527b3114aeb1a00689cce17498a21ba99e7 [formerly 419bfb4b26fb05725f100c111e81fd2dd4dabe34]
Former-commit-id: 265b128ed9281601d494d487af971612aed9b3a0