wmllint: remove -p --progress option
This option is pretty similar in effect to -v -v. Particularly now that
some new switches have been added recently, the clutter makes it no
longer worth keeping around.

A comment notes the option's disappearance, in case anyone wonders what
happened to it or how to get the same functionality.


Former-commit-id: a502f166e40590b84cb78bbcf0f9604ca0c7569f
Former-commit-id: 64f75a467275e51261d0162ee42810a0898c00a9
Former-commit-id: 1ecc96acf4689f9bac89c3d143c920cc869fa235
Former-commit-id: 23452555d9418f7e11ec5ba23734badbf78e9784 [formerly 1eddabc1febe8498c227c46a0336845c6c438fa0]
Former-commit-id: 01ad3a6743c8d9cd04d7a5f8b10c54b9146108fd