wmllint: remove -p --progress option
This option is pretty similar in effect to -v -v. Particularly now that
some new switches have been added recently, the clutter makes it no
longer worth keeping around.

A comment notes the option's disappearance, in case anyone wonders what
happened to it or how to get the same functionality.


Former-commit-id: d9f4b3991f0949ca69c6d0be19398d9a349909a0
Former-commit-id: 4926b6527b53e5e3545fe411179537376d91737a
Former-commit-id: 1a65ef29a28d154492d9008152a7270fc92a2d8e
Former-commit-id: 551071ba39ea6dcbb520daa9b11693c0a651c07d [formerly 30a4fd295f86ba9cd0f57f22d8e4fe5922acd61d]
Former-commit-id: e9bf82fa4e51ebb6efdf817d1266e80970462594