wmltools: split campaigns/add-ons into separate subtrees for wmlscope
The problem: wmlscope assigns each file to a directory-based subtree, and
assumes that it is visible to every other file in that subtree. The intent
is that each campaign (or add-on of another type) will be in its own
subtree.

However, wmlscope/wmltools has derived these subtrees based on what
arguments get passed to wmlscope, rather than checking whether campaigns
have actually been split out. For instance, invoking wmlscope with
"../data/core ../data/campaigns <userdata>/data/add-ons" will cause
wmlscope to see three subtrees, one of which consists of all mainline
campaigns and another that includes all the user's add-ons. This leads to
many spurious "more than one definition/resource is visible here" errors.

A wildcard like "data/campaigns/*" would be expanded to all the individual
campaign directories. However, the wmlscope user might not know that they
should do that. Also, until the glob module was imported recently, globbing
did not work in the Windows cmd shell.

The solution: declare "campaigns" and "add-ons" to be roots, and check for
their presence in the directory names. If indicated, split out the
subdirectories into their own subtrees. Since it is possible that the user
may have moved, copied, or drafted campaigns/add-ons to another folder
that isn't following Wesnoth convention, also check for that.


Former-commit-id: 6178c309ca5fc21b6077883a09213731fcf19804
Former-commit-id: e0965dd277c9a9bbf96216db7fa9e081e650a5df
Former-commit-id: b7539c91ee86277eecfd1860f34ca444795f3f57
Former-commit-id: 7278fc9f6af75b79752e0ee0f74822fa43dababd [formerly 6a666fabe79e8830a940e0931ab398771231978d]
Former-commit-id: d26419ef37b3860a800bfbfb57e62783690bb896