wmltools: split campaigns/add-ons into separate subtrees for wmlscope
The problem: wmlscope assigns each file to a directory-based subtree, and
assumes that it is visible to every other file in that subtree. The intent
is that each campaign (or add-on of another type) will be in its own
subtree.

However, wmlscope/wmltools has derived these subtrees based on what
arguments get passed to wmlscope, rather than checking whether campaigns
have actually been split out. For instance, invoking wmlscope with
"../data/core ../data/campaigns <userdata>/data/add-ons" will cause
wmlscope to see three subtrees, one of which consists of all mainline
campaigns and another that includes all the user's add-ons. This leads to
many spurious "more than one definition/resource is visible here" errors.

A wildcard like "data/campaigns/*" would be expanded to all the individual
campaign directories. However, the wmlscope user might not know that they
should do that. Also, until the glob module was imported recently, globbing
did not work in the Windows cmd shell.

The solution: declare "campaigns" and "add-ons" to be roots, and check for
their presence in the directory names. If indicated, split out the
subdirectories into their own subtrees. Since it is possible that the user
may have moved, copied, or drafted campaigns/add-ons to another folder
that isn't following Wesnoth convention, also check for that.


Former-commit-id: a8ce8ada11480704c91bbd38ebbbe6669586d3d6
Former-commit-id: 0aad7edcb4fc6cad86d324d53cbfd13f6fd84c1b
Former-commit-id: a5d51c1fb188fa1270129a7af096270678cdba2a
Former-commit-id: f609291851bb79f59c3c141e7caad5d4d91b5cb6 [formerly b33a1e2c3feec8059c04962d216aff370fdba1c4]
Former-commit-id: b0fb8d030bb791020e2c6e8903e09454f854c74f