add application lua kernel and game launcher lua wrapper object
The application lua kernel is meant to interpret scripts to drive
the client. Its main features are, it stores a script in its
registry and provides a C++ function to call it with a config
as its argument. Its initialization also creates a "game_launcher"
object, which stands as a limited lua proxy for the C++
game_launcher object.

The game_launcher object is a table, holding a pointer to the
actual game_launcher it represents, and with metamethods to return
version info and the command line arguments. It also holds
callbacks to set the script for the current application lua kernel,
and to call the "play_multiplayer" function of game_launcher.
Currently it is readonly and you cannot write to its fields,
although that may change to allow the target server to be reset,
etc.

Some minor changes that are included in this commit:
- the actual C++ game launcher object provides an accessor to the
command line options.
- the scripting/lua_types.hpp file now has include guards


Former-commit-id: 5608f472cb7c5123a24e4f31fb3683bccb738615
Former-commit-id: d9549b90d7e41df4e0fae24b45e75a45d62fe152
Former-commit-id: c0cbe2ccdde7e69c84d4697d6d14d62ebbf9cae4
Former-commit-id: 5adb5a7e8bd47d767a554eb0a3b38807b565e57e [formerly af4a252552768323b4fbf737f697d25e0093a6e6]
Former-commit-id: 80c62ba088096983b567cd1984871464b87f8567