make "unit has empty type" error not fatal when starting a game
This can happen easily in a campaign if there are saveid mismatches
and such, and generally the only error message will be "game error:
created a unit with empty type". This commit makes the error not
fatal, and simply skips the unit. Also we give a better error
message, on the warning channel, which tells the id of the unit
and the details of the side which owned it.


Former-commit-id: fe05ff0594f5393bd53a879bdb943a4d5ce8297e
Former-commit-id: 29d9e8780c035577ddf60d0a8fa8f2da6a1ff3f3
Former-commit-id: dd63b033348e4d3cb1eb19bf3cfbf8d3c068ac9b
Former-commit-id: a1028314e2af8f4ca4617bae5deacf3a625d6804 [formerly df7e832cfac1c769e81c24b10864ccb0067ab858]
Former-commit-id: 8ce078826c02b730639d5f8fda9e9b07c9507c01