fixbug 22611: don't close the program if we can't find button image
If the main button image could not be loaded, the previous code
would throw an exception of an obscure type "error". It appears that
this was not caught anywhere except the main function in wesnoth.cpp
and so the program would close immediately.

I first tried to go into the gui initialization and construct
buttons in try catch blocks, skipping ones that have problems, but
rather than being able to proceed, this just causes a segfault
because other parts of the code assume that all buttons have been
constructed successfully. This means that not finding the image is
necessarily a fatal error.

My fix for the problem is, instead of throwing error, we throw the
more standard "game::game_error". The new behavior is, if you try
to load a theme with nonexistant images, then a black screen will
appear with a dialog

"error initializing button images! filename: buttons/dontexist.png"

The game then returns to the title screen.


Former-commit-id: 71be3cfd6ba6018ac482df6c7574fbf27b07c344
Former-commit-id: 500f75cfef19b1af439405c545e3e59e86c918f2
Former-commit-id: 6382fd7d4724b42be6e0f06bc7171a16c3b1b0c4
Former-commit-id: 4af13646772417524ceedc98d36a7ae2f8584bdb [formerly dc58222a2075b330f29abf00a0ef607b72c21a99]
Former-commit-id: 95c1989c73f9ab3e3907c32c4f66960820251811