fixbug 22611: don't close the program if we can't find button image
If the main button image could not be loaded, the previous code
would throw an exception of an obscure type "error". It appears that
this was not caught anywhere except the main function in wesnoth.cpp
and so the program would close immediately.

I first tried to go into the gui initialization and construct
buttons in try catch blocks, skipping ones that have problems, but
rather than being able to proceed, this just causes a segfault
because other parts of the code assume that all buttons have been
constructed successfully. This means that not finding the image is
necessarily a fatal error.

My fix for the problem is, instead of throwing error, we throw the
more standard "game::game_error". The new behavior is, if you try
to load a theme with nonexistant images, then a black screen will
appear with a dialog

"error initializing button images! filename: buttons/dontexist.png"

The game then returns to the title screen.

Conflicts:
	src/widgets/button.cpp
        (just the #include reorder)

Former-commit-id: b4f21aac08f707977fd41cc3d554e75723fa3bdf
Former-commit-id: e2fcf33bf71680942a691e273ab44357a6fd22fb
Former-commit-id: cfe092668aacc3b4402acf1c05e7c7ded63518b1
Former-commit-id: abd3506faa1aa8fe9f2439babbbf2edc30ab9a13 [formerly 7bf1f3047a35a4cb44fc0c11ebacd2c5d366b848]
Former-commit-id: 3be4edfad7046161fa9e58bcab1df0331a1fa9c1