add try_fetch_addon_with_checks fcn, use when gui installs addons
Extending the 'try_fetch_addon' fcn, this fcn also calls the dep
check and the check for dev overwrite functions beforehand. Thus it
does a fully 'smart' download of an add-on and tidies up many of the
details.

We plan to call this function also as part of the code path when we
do addon auto-downloads from the mp lobby.


Former-commit-id: 0e995b5501d7f5e9de7e5931b5e50af660bc670d
Former-commit-id: 75c0f251e8fe46db2f112ffbf822115b3de2f9f7
Former-commit-id: 51d734bdebed65aa7b82fe4b66891d3f07bbbbee
Former-commit-id: d0931105c9bc9d37713c9656b7d0b93c1837aa06 [formerly 6e48fd9916d2d7b6b62c0a685db4127526479fe6]
Former-commit-id: cd371e4541085b9a36a4636f14e379b134a1dc39