add ad_hoc_addon_fetch_session fcn, installs an add-on given its id
This function is public and placed in the header, so that it may be
called by the mp lobby code.

Essentially, it just takes the client creation and connection
routines from addon_manager_ui, calls the try_fetch_addon_with_checks
function, and wraps this in the exception handlers from
addon_manager_ui.


Former-commit-id: dbcccc75fe02af119b96fd6e96268483178ce0eb
Former-commit-id: 8d2703f235ae1d84c607bab7807aaaa8f8e8adfe
Former-commit-id: 1c9e135d0e6855be73f9340e1d07b65d818e80e3
Former-commit-id: dcc2055e6820738a786205c11a54d901515f36aa [formerly c6c9f21361639ca79d06ef9d91dc09da9128bfd8]
Former-commit-id: 1d5923a74ab6b80e40e986210338858ab10fcfc8