in game cfg mgr, annotate era, mod tags with the addon_id
This is done at the time of loading the config of all addons. The
addon_id can be deduced without much difficulty from the path, we
do this calculation, then scan over the entire cfg that resulted
from that addon and add an addon_id attribute to appropriate top-
level tags.

The purpose of this is so that this info will be included in a
level config after it is constructed by the host, and can be made
available by connect engine to clients that wish to join.

It might be better to do this even earlier in the pipeline, i.e.
so that it happens before the contents are cached. At the moment
I don't see a compelling reason to do that though, and this way
was fairly simple.


Former-commit-id: 66ff01ebd79707c0627da8e6d8edcb507c8c9ad4
Former-commit-id: 1823c9bb11e632157b408036909482c72577c87a
Former-commit-id: bc45bcdd2981d1ba3f3d879e2d93f85473d4bd1f
Former-commit-id: 181495eb4e5a5374c4f983ad98bbb41978b485a4 [formerly e3b46088b7c65990d673e995b43fc6846f8326c5]
Former-commit-id: b8b1524a836604618c22f57d48df7b8c872606c9