in fcn expand_mp_events, also collect addon_ids, for mp_settings
At the same time that we expand multiplayer events, by looking up
the tags of the given ids from the game config, we also note their
addon_id field if present and write this list to the mp_settings,
which will become an attribute of the [multiplayer] tag for the
scenario. When the scenario is hosted, this will tell other
clients what addons they can download to be able to play.


Former-commit-id: acf9124ef7fe3b070801ae8daf5d8f1870cfe7e8
Former-commit-id: d16b8856eaffcc68b3a9d5e2f34a651c172fc65d
Former-commit-id: e36dba499a11d5c916c878f23ec1e9608dc32024
Former-commit-id: 9fc3cc2798ff5f7269792a16878fcd3290adc380 [formerly e1a9ace565d44ab5237326bf465082e317e466ee]
Former-commit-id: 01b57d47abecd572c1c60fc941984e133f37f397