in mp lobby, try to download addons if we can't join a game
If we try to join or observe a game but don't have addons to do it,
ask if we want to try to download them, then call addon client code
to initiate transactions.

Note that because the "is_joinable" / "is_observable" logic includes
this stuff, we would have to refactor to be able to determine
"would this game be joinable if I downloaded the addons". So in the
current state, it's possible that you will be prompted to download
add-ons for a game that you ultimately cannot join anyways, which
might be a bit dissappointing.


Former-commit-id: 1c9392851ee166c9b1d14b0bdae64de33213ba4f
Former-commit-id: b8b2399568d855305bce0049568cb8ff724ee95a
Former-commit-id: 81492c321a518f9a9b969baebef5a6747c29958e
Former-commit-id: 6d51590c2b26d91b251110df9cd2dda5ae8ef3d9 [formerly c818ca17bbe8de30b43f3ef47e1dbb7674f0d48f]
Former-commit-id: f0625ca4d376faa4b91a2f3ca15f5a92995c2ae1