disable "alpha thresholding" in our bilinear interpolation algo
In this commit from late 2005, boucman added a bilinear interpolation
routine for scaling images which we use everywhere.

https://github.com/wesnoth/wesnoth/commit/2df60619

It is a modified bilinear interpolation -- one of the tweaks is that
at the very end an "alpha thresholding" step occurs. If the alpha
of a pixel is less than half of the average alpha of its neighbors,
it is discarded and set to fully transparent. Boucman writes that
this was done to combat a gametime artifact that would occur when
moving the hexagonal selector around with the mouse -- however, it
creates a small loss in image quality. Boucman writes that skeleton
archer bowstrings don't look as good -- I notice also that the edges
of the wings of drakes seem a bit sharper / more pixelated somehow.

In this commit I tried disabling the alpha thresholding, since the
way that the hex selector works has been changed since 2005, so this
might no longer be an issue. Sure enough the "hexagonal haloes" are
gone now, so it seems that we can safely get rid of the alpha
thresholding.


Former-commit-id: ab18f52c3f9e68ff08fe56a46d568afba5cea80b
Former-commit-id: 91a380cc9e2ed2670f893e51b8dabf93a2214198
Former-commit-id: b42b2ade436996f5e1c2ef3148b8d6560ad58c32
Former-commit-id: 1abac779bd447cc934f1c9af06e7acd3e62cb968 [formerly 7efb7f2c11ee482b784fab7269dd70a07d00c8e0]
Former-commit-id: e435acb508b9c858819c7bd278fab445132826bf