disable "alpha thresholding" in our bilinear interpolation algo
In this commit from late 2005, boucman added a bilinear interpolation
routine for scaling images which we use everywhere.

https://github.com/wesnoth/wesnoth/commit/2df60619

It is a modified bilinear interpolation -- one of the tweaks is that
at the very end an "alpha thresholding" step occurs. If the alpha
of a pixel is less than half of the average alpha of its neighbors,
it is discarded and set to fully transparent. Boucman writes that
this was done to combat a gametime artifact that would occur when
moving the hexagonal selector around with the mouse -- however, it
creates a small loss in image quality. Boucman writes that skeleton
archer bowstrings don't look as good -- I notice also that the edges
of the wings of drakes seem a bit sharper / more pixelated somehow.

In this commit I tried disabling the alpha thresholding, since the
way that the hex selector works has been changed since 2005, so this
might no longer be an issue. Sure enough the "hexagonal haloes" are
gone now, so it seems that we can safely get rid of the alpha
thresholding.


Former-commit-id: 96ccafea82751cb8e9cfc7613843e0a8d075ed8d
Former-commit-id: 941476784b0c358705d1f8c8128f0cc349e65248
Former-commit-id: 9953fbaad9c26aeaf39d407eaaf835fb10cb0c8c
Former-commit-id: 925a7d14d33b65b7ffe7d8adf4ee2dd41be057cc [formerly 0410ce53f52a6fecabdcc468ae85a41ce82b8b24]
Former-commit-id: 56534abc3e016c644440195812c50b7d0c032910