Detect empty subnamespaces in [clear_global_variable] (bug #21093)
Once they become empty, there is no config left behind to clear, which
causes persist_context::get_node() to consistently return NULL.
Previously, this caused a straight null pointer dereference; in order to
guard against undefined behavior, an assertion check was added to
properly terminate execution first.

Neither solution is correct. My understanding of this unwiedly mess is
that persist_file_context::clear_var() recurses outwards from the
original target pruning empty subnamespaces. At some point we are going
to run out of things to delete, so just quit the recursion at that point
and return to the caller and everything should be fine (or so it was in
my test runs with different subnamespace layouts).

(In general, the fact that this code completely lacks documentation of
any sort and has several non-obvious layers to it is a *really* bad sign
for maintainability of this feature in the future. Furthermore, this is
not the first time we have had problems with glogal variables nested in
subnamespaces -- see bug #20385, which was arguably more disastrous in
practice.)


Former-commit-id: b60357a1d0c7c7be2a79e58ad855b6cd14ab7b57
Former-commit-id: ef1a741f3002edb3057020ad3e665315936534a7
Former-commit-id: bc9c2098c754fa1f1a94430dc0a47cc0131883be
Former-commit-id: 8a9bb4efa8d93e20ae6409e72888c27be472e9f5 [formerly 251865a0163a107e7cf051253b4ba1113f0d6c96]
Former-commit-id: ec393f3c99aa54044baf2ed4363c5386f427645e