Detect empty subnamespaces in [clear_global_variable] (bug #21093)
Once they become empty, there is no config left behind to clear, which
causes persist_context::get_node() to consistently return NULL.
Previously, this caused a straight null pointer dereference; in order to
guard against undefined behavior, an assertion check was added to
properly terminate execution first.

Neither solution is correct. My understanding of this unwiedly mess is
that persist_file_context::clear_var() recurses outwards from the
original target pruning empty subnamespaces. At some point we are going
to run out of things to delete, so just quit the recursion at that point
and return to the caller and everything should be fine (or so it was in
my test runs with different subnamespace layouts).

(In general, the fact that this code completely lacks documentation of
any sort and has several non-obvious layers to it is a *really* bad sign
for maintainability of this feature in the future. Furthermore, this is
not the first time we have had problems with glogal variables nested in
subnamespaces -- see bug #20385, which was arguably more disastrous in
practice.)


Former-commit-id: 0edefc2966b210dcc9d17f095dd350a8d6cdf564
Former-commit-id: 24d90b8db7aa79d6e467d502eaeac6f7a4326dff
Former-commit-id: e6b27cb70f8ad7de48527ec8528685ba71d65540
Former-commit-id: 83fc874f21e43c340ffa4a5fafcf84d18b532fbb [formerly 1ba70ab2dfb2cbf42340ad78ac54620ca639f0d8]
Former-commit-id: 0118095cc6e379a4b63a82eda8c142d20b758a97