campaignd: Add server* this parameter to server::request_handler type
This gets rid of the need for binding this pointers for every registered
request handler in the handler table, and saves us some negligible
theoretical overhead from boost::bind. As a bonus, the
REGISTER_CAMPAIGND_HANDLER macro is more readable.

Now we pass the current pointer during the actual handler calls instead.


Former-commit-id: e3928ff031a61f82964a5f8eb7f75555b1310401
Former-commit-id: 45fc309b5d07bad94a4f74d8947ac3d5effbf3d1
Former-commit-id: e890e740c06935f73b6788d0546b81d8723b0c82
Former-commit-id: 2e2fd70a0e39c8f830f82bf7ffc85febd63357da [formerly f9c3d55e3c725f045b1019bffc56f486c7b5db88]
Former-commit-id: f3f7de4ba5ccfd8de22617d166f18f220638f271