make all lobbies, mp ui screens, and chat handlers use the new code
This commit replaces all appearance of `game_config::sounds::...`
where ... is an mp ui specific sound, with an appropriate call to
the mp_ui_sounds code. This makes sure that the preferences are
checked appropriately and notifications are distributed in all cases
appropriate.

This also obsoletes the "lobby sounds" advanced preference, it
doesn't do anything anymore. Its removal is pending.


Former-commit-id: e12c58c968c5974667104d50fdf496dc3d34a883
Former-commit-id: 7c69cbc58ea423ebe5252e7ef7c00eeef86c3b16
Former-commit-id: abf1a66eec3786199220f1685437ed18e8e9ee7c
Former-commit-id: c6e2e0c7a1a2dcf7114df94e838deea544c31aab [formerly 9d09e1aa6a37c4e2be312f66d5f3fa976de480f0]
Former-commit-id: 6bfd3eb1f0314638c5013d04a7b7c74345541e0b