fix bug where preferences seemed not to be saved
The problem was that widget ids were being used in place of
preference ids, which are very similar but not exactly the same.
This commit fixes the bug, but goes further and changes all the
widget ids to match the preference ids, reducing the chance of
confusion in the future.


Former-commit-id: 825bde328ff23de3226e1b83d11ee967e732305c
Former-commit-id: b818c6f6d9cd66d64c6ad2ab588c195b52dd2911
Former-commit-id: 4b8d480c19cec2f5d662a83d45567b99208454cb
Former-commit-id: dd2858a0986e433f2af810023442c6f65e306ceb [formerly 7c0cf5108393878579a98a04c6cef594acdfe47a]
Former-commit-id: 2a2d8d540915877994ee4d58f4a2af163ab428c5