fix bug 21400: unit::absolute_image being used for unit animations
absolute_image defaults to be the image_icon of a unit, it is
properly used in the sidebar or in relation to portraits of a unit
it only takes the value of "image" when there is no image_icon.

prior to this commit, absolute_image was being used a default
value for animation frames. for many units this is okay, and results
in the base frame being used, but in the cases highlighted by
doofus1 it's broken and results in a portrait being blitted onto
the game map, when image would be a better alternative. This commit
replaces the uses of this in the unit, with either the unit's image
or image_icon, but prioritizing image if both are available.


Former-commit-id: 476365f1754b749d6e53cba6fa5ae9863e43e00b
Former-commit-id: ec70d35b4997249c836b02e4ead223daac63fc0e
Former-commit-id: c8301ad83568859e1e80833b353b17bdfe0e3603
Former-commit-id: 77cd78b444900eb8b1b891c597d51a7b8aa958f3 [formerly c3edabc08cbe586e7f6728ae9ac4aac0cb8b12b9]
Former-commit-id: d432c7edc863f679b594747196510333acc6bf76