replace "get_string_index" with "get_token"
The get_token function, for a font::ttext, takes a coordinate
position and a list of delimiter characters, and returns the token
pointed to, if the position is in bounds, and the character being
pointed at is not itself a delimiter. Otherwise it returns the
empty string.

We also add some debugging output to the label click handler,
so that on log channel gui/event, we can see example tokens and
whether they look like URLs.


Former-commit-id: 49cc55beffb70b52b36ae39912d7fcbfa098d127
Former-commit-id: d7fc09308a80471c5fc901a60446bf49589e3daf
Former-commit-id: 33de553d2c79d151c256a3b24584dc99afcbfdf1
Former-commit-id: d2dc0aa781440768cdea2197a93b2f25aba20795 [formerly c32997c9805fcd0a4209d1fb4fd04d71739a273d]
Former-commit-id: 4584f2ce806afb39ca30afd27d917cce14c60769