link-aware rendering of labels
This commit adds a new property of font::ttext, whether it is
"link aware". Link aware text has tokens that look like hyperlinks
formatted thusly, with coloration and an underline.

A `get_link` function, taking a layout location, returns a link
token if it is pointed to at that location, and link awareness
is enabled, and the empty string otherwise.

All gui2::tcontrols may now have their text labels set to be link
aware if desired. The gui2::tlabel constructor automatically sets
these to be link aware. The gui2::tlabel click handler is set to
use the `get_link` function instead of the `get_token` function.

Link awareness is also registered as a property of [text] fields in
gui widget definitions.


Former-commit-id: 09a6c79f8883730e5b005a9f4867572dad9caeb7
Former-commit-id: 0681a0fcd22d13113d8cd4f7361d33eb13dd8cc1
Former-commit-id: f185c4fff2e72b674e65a6edffb5a9622565e90f
Former-commit-id: 7fe462f899db38128912d2a26a78181992f916b4 [formerly 45d864f5db574b6425d2951c0c1bf12254861bfc]
Former-commit-id: cb3ff7a18ce44c7dc8d8e054942aba104ab61031