add link_color property of label defn, passed to font::ttext
This commit permits link_color to be specified in any label defn,
at the same place that link_aware is specified. It will be passed
through to tcanvas and finally font::ttext in a similar manner as
link awareness, and used for the color parameter in any link
formatting. We choose a default value of #ffff00 everywhere,
matching the definition currently in the help browser for cross-
references.


Former-commit-id: 27b00e8b278ef25bd345d2a8c061904f42fd125b
Former-commit-id: a39a380c4d1aca7d5a7b13388fa3f2a1607a8b05
Former-commit-id: 03f3eec7fea12dead6e664cbb4de7b9e7d39a47d
Former-commit-id: 232b85c87a03e2e28941e41fc3549302f0d96f00 [formerly 86b74bdf1bea2bbef6d3d3fea2d73cd2a0ee3b23]
Former-commit-id: 413dc7ba56fcd5b968424c310230c9bc5c51dc17