Merge branch 'boost_filesystem' of https://github.com/AI0867/wesnoth into AI-boost_filesystem_3
Conflicts:
	.travis.yml
	SConstruct
	src/campaign_server/campaign_server.cpp
	src/create_engine.cpp
	src/editor/map/map_context.cpp
	src/game_config_manager.cpp
	src/gui/dialogs/screenshot_notification.cpp
	src/gui/dialogs/wml_error.cpp
	src/savegame.cpp

Conflicts:
	.travis.yml
	SConstruct
	src/save_index.cpp

The second conflict list is from the cherry-pick, the
first is from gfgtdf's original merge.

For me there was additional conflict in
filesystem_boost.cpp, because the
"treat_failure_as_error" flag does not exist in 1.12
branch, it was introduced to accomodate unit tests.

Additionally, dirsize was added even though its not
needed in 1.12 (I removed it), and the change

remove -> filesystem::delete_file

which appear in save_index.cpp in master needed to be
recreated in savegame.cpp in 1.12.


Former-commit-id: 0f7e7f9ae56a98f9b466d134440e5eb096d4903d
Former-commit-id: b0cd2bee891082de8f98a8fc87607fec765c6b92
Former-commit-id: dcf7fa6dc0a7c76b6044d4ca37701a01651716df
Former-commit-id: 8e526387e1c11a21904833e093b52f24eddadc89 [formerly 3fe788dbe1ef22d940897f75d788509f78b19223]
Former-commit-id: d396a977ce454f328df1b49bdf5c53a6501659c0