add exception handlers for "game::error" in playcampaign.cpp
I also add one for mapgen_exception, but this doesn't make so much
difference. It seems to be a big problem that we don't catch
game::error anywhere, I observed a similar bug being caused by
game::error being thrown directly by a button when it couldn't
find its image. The report was that it caused the game to crash
to desktop immediately (I think it was caught only in void main.)

Former-commit-id: 6a58472a17373bbf04dd7bdb89d5c6474630f23a
Former-commit-id: 31a0f9e3d645b286fe23a780d2ebc4eec262259e
Former-commit-id: 2f46de04bf1e01223727e12349876292da1a2d01
Former-commit-id: f7529c2f24662207034624e5bce8460f4d6c427b [formerly b0a0a9a677f2eddedff1d4873b76ca50b8d3e74c]
Former-commit-id: 0e8a29776955b09b6e9c5c572f8bb2ce1d9267b8