fixup xbrz blending handling of alpha channel
xbrz reference implementation assumes that there is no alpha in
the sprites, and indeed, when it would blend any colors would zero
out the alpha channel. This is not suitable for our needs, as our
sprites have alpha, and SDL treats 0 alpha as being transparent.

This commit adds --placeholder-- alpha channel code. It does not
do proper alpha compositing, it just checks if either the source
or blend pixel has < 50% alpha (if so it ignores and takes the
other), and otherwise it applies full alpha to the pixel. This
seems to work fine in the ``typical'' cases I have observed but
a ``real'' implementation is needed here.


Former-commit-id: 05b601643afae69f2ae7d7a15cfa1309fac2145a
Former-commit-id: e82752e3cac0d2cbe4662401b7953d9c4fd7ba32
Former-commit-id: a128ac6ccc5eedc98207d4e842b43750392396dd
Former-commit-id: efad81792d02f43702b5d6f3dd050f2bfd921084 [formerly 1be34aade552e29ce045c4b3109666232345b2fa]
Former-commit-id: eda039e152510da8677a9074757d75e6a748ebdf