add "advanced graphics" preference dialog, for scaling algo options
This commit also gives the smart enums in src/make_enum.hpp
introspective capabilities, by giving them a static const size_t
which tells how many possible values they may take (hence the range
of the enum.) It was convenient for the dialog implementation
to have this.

Note that, I had to make use of __attribute__((unused)) to compile
this. I apply this using a macro which adds it for gcc, clang, and
mingw which I know respond to this attribute. I don't know what is
the equivalent MSVC device, or if it is even necessary, so its
possible that this commit needs to be fixed up.

We also disable unit tests for the new advanced graphics dialog,
there are no tests for it right now...


Former-commit-id: e001581f8fa8b3c4f42461fb3e3dd0cd532af778
Former-commit-id: 844088322861592b390781f4b28b2682a4b5eae4
Former-commit-id: aad0c36d3bfedc19ad8cba53bd7d8133babac104
Former-commit-id: d3ca8be6863877ee51861f5dc23c3d1c9d60b414 [formerly 52ec094826e4aa4a59d24efb9561871d8865db89]
Former-commit-id: 7b185693491fd82031a67f58c5b9f23fd1977f59