disable auto-backtrace for C++ unit tests, try to get stderr again
We have some problems where the test executable is not generating
all the right output on travis, it seems not to be getting stdout
or stderr or something. The test executable was being run through
gdb at all times, with commands to backtrace on case of crash.

I'm not sure if gdb is eating the output or what, but for now I
am disabling the feature in the hopes to get the error messages
again.


Former-commit-id: 7d9c4fbdd39d2a5ab3890b0d55aa39dd0ccaf515
Former-commit-id: b26ee5d060299418e68bac20dac3faaf90e1cce8
Former-commit-id: b590177049deba9f5e5e2a727dc59839b06a5eec
Former-commit-id: cbf40a4c6eb9008946eaf2685c71c5612e2a1b70 [formerly 17a708046848a450ff242985183d377571a4f525]
Former-commit-id: 636d5386b376c1be79faec7d942f6c35f740eac7