add SDL_SavePNG, write files as png if possible, unless ending .bmp
This commit adds a "minimal interface to lib png to support writing
png files from SDL surfaces", to the codebase. It adds a save_image
function to namespace image, and this checks compiler flags to see
if lib png is available. The function will save as a png if png
support is included and save file name doesn't have bmp extension,
and otherwise it will save as a BMP as usual.

We already had lib png as a dependency to build the tools, just not
the main game. This commit changes cmake and scons but makes the
dependency only optional, changing nothing if lib png is not
available.

See here: https://github.com/driedfruit/SDL_SavePNG


Former-commit-id: baba6099d7737edef75f88e14eada31023873ab3
Former-commit-id: 4def23fdda598ab6d31d2d93a792ddb363559063
Former-commit-id: 8c5864245578aca25aac2849eeccba029c3c89ce
Former-commit-id: 28637a4038359f1872aaaef3d58718aaa565f190 [formerly 026a4e5c2fbffae816e984a838c1ed6e17b15370]
Former-commit-id: e177275145d96d150bfdde503acb214b153eabc6