Xcode project: update boost and SDL libraries
This updates to using boost 1.54 (also adding a new library:
libboost_filesystemw.dylib) and SDL_mixer 1.2.12 (plus the other SDL
frameworks, but their upgrade is not essential), in order to work with
the new filesystem code that will be implemented shortly.
This also means that a new Mac compile package is needed.


Former-commit-id: 495e1b2ff2b764e0e62d8c14d6e845a9ed21da36
Former-commit-id: 980c4135c9e6a2854f53fde79eb885659615c0d5
Former-commit-id: 7aa13f6715e2bf1c897830623c3ab13f25ab1922
Former-commit-id: f3a030d0a8eca35a71c98552e36e2296d4bbaf6e [formerly 36d99398771b38fceec03c4c0cbf1e9ddb209caa]
Former-commit-id: c10c8aed6e26928e317784123cc869f2d3ea7bfd