New Ancient Lich baseframe and updated animations by shadowm
This version of the commit removes the orphaned melee attack animation
frame, which is superseded by a 2-frame set. The 1.12.x version of this
commit keeps the file as a copy of the new melee attack animation frame
for compatibility with existing add-ons.

Animation WML and art reference taken from the existing L3 Lich.


Former-commit-id: 047df6537687675974e26f7fdf6e2c220d9505f0
Former-commit-id: 75a82f5dd4a69ce2633bd30587a40b9a087ada94
Former-commit-id: 8f0db242534e33b1bfaca573c4f2feaf77967da1
Former-commit-id: dd8d8d86319bac31730988a7ce139c1461dd9601 [formerly 9506f11b81896cf368575f5ae7d1b46c09db55bc]
Former-commit-id: 0a7d26d801b2bd709fef43c14432b9020624f50e