New Ancient Lich baseframe and updated animations by shadowm
This version of the commit keeps the orphaned melee attack animation
frame, which is superseded by a 2-frame set. The orphaned frame is a
copy of the new first frame, for compatibility with existing add-ons.
The 1.13.x version of this commit removes it entirely.

Animation WML and art reference taken from the existing L3 Lich.


Former-commit-id: 13b2b8d16445e6475a415a588dfdec251efefbe4
Former-commit-id: 3160214ed50fc8b8197f6ce6b3b857a903b0cb74
Former-commit-id: ee01d356fcebaf0731d6c8142a7916f4af4638df
Former-commit-id: dea249bcf37d28ffb03046ce79423fa1fe069f90 [formerly f24685b3ba9341ebdc22e9b4c2319fbd0bd270ad]
Former-commit-id: 49c2481a9d1a8e2ca883f4de194b1f441956f2c5