use boost locale gettext
This includuces an alternative implementation of gettext.cpp. Libintl gettext doesn't seem to support utf8 filenames on windows. Also it causes problems when changign the language on my msvc build. That why i want to use boost locale gettext at least for our windows builds.
This requires boost locale which is introduced in boost 1.48 (That's why i still hesitate to make it the default on all systems).


Former-commit-id: 7d706a0cadb2843bad42a945fcd50a3e23f4bc65
Former-commit-id: 25a51fd9d05deb704a1af552ea73c0572fdc14c5
Former-commit-id: 5895fcd5120e3c5779a9e10f428ec90c407c2b57
Former-commit-id: 681dd86be3153d9539c2ca5b69957d962c7fd11d [formerly 05d29c936748295bbdee3005917ede88c789ca77]
Former-commit-id: c8929d2241964c23db6bf7c28a2c01b5014f0d70