backport our unicode functionality
This commit backports the src/serialization/unicode files.
I'm not sure exactly why they weren't backported before, probably
it was thought not necessary. However it will be hard to backport
gfgtdf's unicode fixes without this, and I don't see that it can
hurt.

There were some other issues relating to unicode that were not
backported in this manner, for instance in master the addon code
uses the unicode functions but in 1.12 it uses some slightly
different code. We should perhaps roll that over as well.


Former-commit-id: 826f13b155157d82cc6d53463d7e53d40bff28b7
Former-commit-id: 50bd03a52933d7f1939a9e69a74f432c22f189ba
Former-commit-id: 54ce592d8538217012b5c05e01504febf077356f
Former-commit-id: 3f99beca686bc7ca43060e1d381aaf0fd5a5e942 [formerly 8252ce0d54241623988bc78044a29bf1c45521b2]
Former-commit-id: f087f800db9dea7b83f8718685474c2700effce0