when zooming in, use scale_surface_sharp for units, terrain
Previously we were using scale surface. The result is that sprites
look blurry, terrain looks blurry, and for fractional zoom values
"hex tearing" artifacts are more visible. It seems that zookeeper
and others also prefer a more traditional, just-give-me-the-pixels
look anyways.

lipk even thought that we had been using scale_surface_sharp on
units "since forever", so here we treat the fact that we weren't
as a bug.


Former-commit-id: 4955e8a0bc557be903b3f9cb853404e92cb8edef
Former-commit-id: 2eaf841655e8607ea6206d4adb4929cba61edd3b
Former-commit-id: 768cc206ea5653ca276885d4037ff200f83d6281
Former-commit-id: 7d3dd132b848a518d6c1c403ba92854812d35e9c [formerly 9d2c767ad89f7891b207b54057b0baac4cab2475]
Former-commit-id: cce703d234ddd9801ae358165d154dcbad079be0