change preprocessor symbols controlling SDL MIXER min version
Previously, the code that requires SDL_mixer v 1.2.12 was guarded
by #ifdef PANDORA, since this was the only platform intended to
use this. However, it turns out that OS X < 10.5 has a hard time
getting this library as well.

Instead of branching by platform, here we just check the version
defind by the SDL_mixer header, and include the code appropriate.

We rely on the scons / cmake / project file to fail any build which
should be done with the recent version.


Former-commit-id: 98cc33489b6d015c638e5dd4db6839767415a070
Former-commit-id: 712122f49dab16e1b117ec344d6ab3262b9af9aa
Former-commit-id: 4863582a7889b5a20170e9d086a2e2f2f284a467
Former-commit-id: 06b2e68edf031baacca38dcfc8708036ddb6ada6 [formerly 1760e4e1917d24535d56c231febf2d619412108b]
Former-commit-id: 12e7538e5b02da20f446940bcd5e28bce96dab24