fixup preprocessor symbols controlling SDL MIXER min version
Previously, the code that requires SDL_mixer v 1.2.12 was guarded
by #ifdef PANDORA, since this was the only platform intended to
use this. However, it turns out that OS X < 10.5 has a hard time
getting this library as well.

Instead of branching by platform, here we just check the version
defind by the SDL_mixer header, and include the code appropriate.

We rely on the scons / cmake / project file to fail any build which
should be done with the recent version.


Former-commit-id: 89ef93c5040037a6f2ac9abdc1ad705d9f2a8a8e
Former-commit-id: 8fe875d2b3d2dd4bc4485a83cae8989c38eb73a0
Former-commit-id: 1f8c97f94e4b46ad11a98f29dadde0bb6d577dda
Former-commit-id: fe04915ebb0e5656a8b5e621547b9902cf714b83 [formerly 4c0a6fb8bf144d440cbf7aecd198c8a564de774d]
Former-commit-id: c42a0817aca2c87bdecdb0ec5b351bcc6b208516