fix bug #22484 (fix random map generation in mp create)
As described in bug report, mp create had a bizarre implementation
for random map vs scenario generation. Scenario generation was called
map generation, and map generation was impossible. We fix it and make
it work like it works in the rest of the game.

At time of writing the wiki describes map generation wml as follows:

```
To use the default [generator] your [scenario] tag must contain one of
the following keys:

scenario_generation=default
map_generation=default

If ‘scenario_generation’ is used, the engine will expect for your entire
[scenario] sub tags to be inside a [scenario] tag inside [generator].
Tags outside of this will be ignored. There may be value in this, but at
this writing, it’s not clear. ‘map_generation=default’ is simpler and
more commonly used. It is also necessary to use this key so that you can
regenerate a map in MP game creation. In its use only generator data is
in the [generator] tag, all other [scenario] data is placed outside of it.
The exception is if you are making an initial MP scenario available in MP
game creation, for this a [scenario] tag must appear inside of
[generator], containing the [scenario] subtags you want to use.
See “data/multiplayer/scenarios/Random_Scenario.cfg” for an example.
```

This commit essentially removes the "exception" pointed out above.
After this, the mp create dialog treats map and scenario generation
both in the "random maps" classification, and it handles them normally,
scenario generation replacing the entire scenario, and map generation
replacing only the map data of the scenario.


Former-commit-id: 5e04f7caa6c11cd79298a48d04d90720e3aca478
Former-commit-id: 39c22fca5c73900dc9a05cadcc93412b18ab3ba5
Former-commit-id: 97d4e1ad9c829a8e7469d049e2f8c667abd2a487
Former-commit-id: 4d62033f7181af5311fd7d351eb3eeb1c17f427c [formerly 9ccce4d2a51b8a178ac95ca56ae92a4762744945]
Former-commit-id: 3da19a1686460f6acad783036da0495605ce3384