cache minimap images in mp create screen
When we are actually in the game, the image.cpp file uses an
elaborate collection of caches to make sure that all images are
cached and reloaded rather than redoing work (even scaling). It's
questionable whether the amount of caching going on there is
appropriate, it seems it might even be faster if we did less. But
one place that we don't do any caching and performance is bad is
in the mp create screen. There was a bug report in 1.12:

https://gna.org/bugs/?func=detailitem&item_id=21801

In this commit, I implement quick and dirty caching for scenario
minimap images. Each scenario record holds a surface (an SDL
reference counted pointer to image data), and when a minimap is
rendered this gets saved, along with the md5 hash of the map that
got rendered. Whenever a minimap is requested, we hash it first
and check if the cache is dirty. If so we recalculate, otherwise
we return the cached image.

The code is also cleaner and a bit less convoluted than the
previous code -- the previous author seemed to think that surfaces
are "object" types and not "pointer" types, that require an
additional layer of memory management with boost scoped pointer
etc. This was in fact redundant.

I don't know if this commit will actually help report 21801,
the profiling seemed to suggest that it's the minimap rendering
code that is actually slow. But it seems that this can't hurt,
and may at least mitigate the slowness, since except for random
maps it means the minimap calculation won't be repeated.


Former-commit-id: 386196bce1741bd45fd1fd3988a04c07bb77aaf4
Former-commit-id: 8c27d0986c631a0d0c1b6b34f1889381ab4376a5
Former-commit-id: fa577e9d7b9107ab0f106593a5088d7e9ca18efd
Former-commit-id: 77620781fddb38161551c62c2b7867cc1878e158 [formerly 70d6a5d6eeaf89b9188449a130320e90ddc1b5ae]
Former-commit-id: d54a01768867c0e1f9326694e24fbdb1aa14519e