split off terrain data caches from the gamemap object
This commit splits some of the functionality of the gamemap object
into a new class, the "terrain_type_data" object. This class is
intended to hold a record of known terrains and the terrain code ->
terrain type map. It handles merging in new terrain definitions,
and it is shared between map objects. It is also cached by the game
config manager, so that we only recalculate this data when we
change configs.

There is in fact a *semantic* change going on here, which is that
when a scenario introduces a new terrain type by merging, this gets
added to the global cache. The alternative is to make a copy of the
global cache, only for the this scenario, as soon as we have to
dirty the cache. However since I didn't see any downside to putting
the merged terrains together and carrying them across scenarios,
I opted not to do that here.


Former-commit-id: 301b3da0d6bd071a51bb6d43ddd0901a0298d0c7
Former-commit-id: 322ea19902d663e3df9c1cebceebf807458c7926
Former-commit-id: c7b415df6582438422c2e2a6de45b12f83dd6769
Former-commit-id: 0c1672f49a7c377132be10667d70e4504256939e [formerly 826a41ced6949c7bf17b0d63bfd5bd5a06590ffc]
Former-commit-id: 3295f3e8a2052cce8f7f960a4ae5499c2cb29315