fs: Fix normalize_path("") incongruity with Boost.filesystem
The non-BFS version of normalize_path() returns an empty string when
passed an empty string, but the BFS version returns the normalized
version of the current working dir for the Wesnoth process.

Due to the way editor::start()'s arguments are built from the process
command line, the result of normalize_path("") gets passed to it when
starting with the --editor switch and no map path argument. If the
result describes a directory, the editor brings up the filechooser
dialog on that path; otherwise, it attempts to open it as a map file.

Making the BFS version of normalize_path() follow the non-BFS behavior
fixes the editor unexpectedly bringing up the filechooser when started
from the command line.


Former-commit-id: 09cc72fade92e2f22cdc5a2c527257f538f0091b
Former-commit-id: 5dd85bc38bf6ff9c3ddf499fd16bfb5ac4fa400b
Former-commit-id: 7e6c867ac229764c554f95a41bfa7a3ad97f1827
Former-commit-id: 0e1c8dacf0fd80d031428f52a518420b9fef496f [formerly e775b7665305bca7f49d8ff95bac8eff50ea9717]
Former-commit-id: cbd7a03100294523595e631afe069c60521a31e7