fs: Fix normalize_path("") incongruity with Boost.filesystem
The non-BFS version of normalize_path() returns an empty string when
passed an empty string, but the BFS version returns the normalized
version of the current working dir for the Wesnoth process.

Due to the way editor::start()'s arguments are built from the process
command line, the result of normalize_path("") gets passed to it when
starting with the --editor switch and no map path argument. If the
result describes a directory, the editor brings up the filechooser
dialog on that path; otherwise, it attempts to open it as a map file.

Making the BFS version of normalize_path() follow the non-BFS behavior
fixes the editor unexpectedly bringing up the filechooser when started
from the command line.


Former-commit-id: cfbd6568c0b7070e3ee8d9c1bdd8106513623c02
Former-commit-id: 1a6471e1a0bfca8bf1f7fd922b9cfbba3d87165e
Former-commit-id: 4a1080f6637b78d6f8d3059af1b41504369a684e
Former-commit-id: 0711ff8f3c44eaed715cd996dd374e3d70bbfafd [formerly 76a728c25aad1a40a8dd949a8bec86e4570c35e5]
Former-commit-id: 35edd86cfc0f8ec814d5ff510409e153cb08fe70