remove comment from dev concerned about a nonissue
read here: http://www.lua.org/manual/5.1/manual.html#lua_CFunction
When the C function returns with "1", all but the top entry are
discarded from the stack. If this function could be called again
and again without ever returning and without ever dumping the stack,
then there could be a stackoverflow, but in this case it's
not possible.


Former-commit-id: 5328eb1f32d34282e2f2d9d7e937defb92737518
Former-commit-id: d1de4a2cbfde66d6cb85e522b4455f248e6d014e
Former-commit-id: e21aee99e33e8ce43bdb7450dfae6e39acc7a937
Former-commit-id: b78e8f4a1e58531a6edc3ede8baa683b29da1537 [formerly e8defa1abe6f4a609a8508ac3c798db06d6170c8]
Former-commit-id: c322f6878ce6cd091c02b245457e0a39f0c4e257