fixup stack discipline with error handling
When a C function terminates, it should not leave bogus elements
on the stack, or the stack can eventually overflow.

This is not as critical for C functions which are called by lua,
because they return an int explaining how many values on the stack
are important. But for C++ facing functions it is very important.


Former-commit-id: b5d0d4c08154d1e0885b40fe8db7909f3d982f04
Former-commit-id: 8e0c26c9e71cfd4217f416d5e1db4d3229df3e7a
Former-commit-id: fa7e7ce8ee7d5b663dcc8bf6ac285e53b41d7112
Former-commit-id: 0750a68970f77f5bd39e7ae39c0d6f27b97b3a23 [formerly 76843928b6a83601261bd4c476b85a85f63400e5]
Former-commit-id: a29bb041b6c3ce634698f1f426c044786e048b98