Revert "disable "alpha thresholding" in our bilinear interpolation algo"
This reverts commit d977314e4d67ff315b0b16f9aa64c0e854a8752f [formerly 56534abc3e016c644440195812c50b7d0c032910] [formerly 925a7d14d33b65b7ffe7d8adf4ee2dd41be057cc [formerly 0410ce53f52a6fecabdcc468ae85a41ce82b8b24]] [formerly 9953fbaad9c26aeaf39d407eaaf835fb10cb0c8c] [formerly 941476784b0c358705d1f8c8128f0cc349e65248] [formerly 96ccafea82751cb8e9cfc7613843e0a8d075ed8d].

This is being reverted because there is still yet more bad behavior
in our alpha channel when scaling images, for instance this causes
the terrain overlays to disappear when mousing over terrains with
one of the editor terrain tools.

The reason is that our "scale image" routine is broken and doesn't
weight color contributions by alpha. It's not apporpriate to get
rid of the hacks like alpha thresholding until that has been fixed
and that can't be fixed on 1.12 right now.


Former-commit-id: e5401c2877ab20c4fa0c6eee46c6a373230bfc0b
Former-commit-id: d8b12dce26de4cd37a681a176936bad144fdf663
Former-commit-id: d206ce06d5d2a02c928b33fafb01f88c4aa8d18e
Former-commit-id: cc0ccdeb447ce4296e6f3639977ef99ade58e573 [formerly 99d6e6829c29bdefc7647df69189b92f5bab3d3a]
Former-commit-id: f7d4a7d01795dbc675c0d70c75d0cdf3e5fc5af9