"Anonymous local player" -> "Anonymous player"
this string is displayed by non hosts too, so "local" is not true since
it's not local from the other clients point of view.
Also the previous string was too large for a unit name in the sidebar.
Note that this does not effect the "Local player" string in the hosts mp
connect controller options.


Former-commit-id: ee03a5507ab611074e8a50a1f4d763f908823475
Former-commit-id: d8b763b053e9c5af3ea00be8392c332e48327f73
Former-commit-id: 6e2bd2f716c34695875e5c01df5e155e9367d878
Former-commit-id: 1ec5f218de8819c32f250dc3880faf1583341bd5 [formerly c25dc56627ebf1398a5e14bc3be18454c85f6a93]
Former-commit-id: ec18b8eafe23e2f6257814879da298e76e03bd45