use sides_.size() instead of MAX_PLAYES
we usualy want to use that number to index the sides_ vector so
comparing against sides_.size() makes more sense.
We now also use side_index for the index of the side in sides_ and
side_num for the value of side_cfg["side"]
I also put some additional checks.


Former-commit-id: e2b3a5f8c09aeaf1aef04ba67f73a4604ee18264
Former-commit-id: e746f4a1b10418a1004fabb23fbcfb1c70105704
Former-commit-id: bdffbbdab091bd850a1e87545c5211742f0b9eee
Former-commit-id: e44743038211b8a5f9969fa510a9feb2196b7b9e [formerly 89224166d0111cd42f46498dc81cb0f1a1f68ad0]
Former-commit-id: 3162013c8bcc3c1370c75188cdaa1950d8550150