refactor lua show_dialog
The lua show dialog used to get the video object from resources,
which is not ideal. In this version, lua kernel base internalizes
a pointer to the video object, and show_dialog is bound to a
method of the lua kernel base.

From now on, lua kernels which want to show dialogs need to be
constructed with a pointer to a CVideo, if it is null the lua will
fail with a runtime error.

The application lua kernel is constructed with the video for
game_launcher, and the game lua kernel is modified so that it is
owned by the play controller and not the game events manager, and
held in a boost scoped pointer rather than manually deleted. This
is because the game events manager doesn't have a video object,
also it conceptually seems independent of the lua kernel anyways.

The map generator lua kernels are constructed with NULL, but if a
user configuration dialog is provided (new feature) then when they
try to run it they will update with the mp_create's video pointer.


Former-commit-id: ed41d87e1e106e82476792e185531ff59899e08d
Former-commit-id: 6291c3cd2aaadd4b0b9fdf311bc8269cd7ef4aa3
Former-commit-id: 5b25ee2897483f9623a0147ef88e68b2e063b29f
Former-commit-id: 0ce5059a9849f505844116bd85ac0ca77869cc36 [formerly 3338682b047746a1f6c4b2fdc143bd37a20f1094]
Former-commit-id: 2e9a8bd98599fcbfcfbb1ca188b127a097bfa97d