fs: Don't complain about failing to read /proc/self/exe with Boost.filesystem
This is intended to silently fail on Unix-like platforms where no procfs
is available, and indeed that's how it works in the non-BFS
implementation of get_exe_dir(). The caller (currently a single function
in game.cpp/wesnoth.cpp) must be able to deal with this situation
in a graceful fashion.


Former-commit-id: 98688daabcb82e39cb8471cf54970b762f3bef97
Former-commit-id: 699a525abf9e3cef840616f74497b20f5b6f9dc2
Former-commit-id: 1fca7553931b90c402e1c4fd9dc38132ff9aa102
Former-commit-id: ffc1ff8bc0092ea0783a82b704d2b1dcaa4169e5 [formerly b2f06ca6dcb8f7c6aca0ff490349b857c2e9e129]
Former-commit-id: 84caea429afba706b21f082427fe478ab9897ac8