fs: Don't complain about failing to read /proc/self/exe with Boost.filesystem
This is intended to silently fail on Unix-like platforms where no procfs
is available, and indeed that's how it works in the non-BFS
implementation of get_exe_dir(). The caller (currently a single function
in game.cpp/wesnoth.cpp) must be able to deal with this situation
in a graceful fashion.


Former-commit-id: 6ff449a3c99788c732d23e17f175cfd98516986d
Former-commit-id: 1a9149e1fa3e12ef507226d3ce293dc89e1f889b
Former-commit-id: fd9fe943e23381fb20f08b926c8c6f08423ced2f
Former-commit-id: 5a700e20143807a441190de14b2666c551e8e4f2 [formerly de085c0ce32da26744f8442db6a03563786e6032]
Former-commit-id: 0398600211aba8e247d816b0b19832dbe420d119