enable "strict mode" in wesnoth lua
Enables an ilua feature called "strict mode" in all of our lua
environments. This change causes lua to report an error if a global
variable is used before it is assigned. The benefits of this are:
- Improved maintainability of lua-based add-ons, since you get
better error reporting in case of a typo.
- Improved behavior of the lua interpreter console, since mistyped
or gibberish lines resolve to an error rather than "nil", which
the console treats as a legal return value.

It is possible to disable this or work around it in individual
scripts. For more info see release notes.


Former-commit-id: 0ab566aa0988d276aa93e407897b150fbafbc320
Former-commit-id: ace25ac6053d061ed4dfd8d9b9d4bd276edc0cdf
Former-commit-id: 149078caad41137cbc07a056f373d720aedddbb5
Former-commit-id: 8b2a8555cf2a0b6a661e5b327ae8210fee296350 [formerly a35011e2f5fbba4c6d37741b8f33a0356e7f315e]
Former-commit-id: 217a4f326c096289ce68b7c28c3b899523d83c2f