fixup lua <-> boost function binding system
The solution in place was working but it was flawed in that unlike
with normal lua callbacks, the callbacks would find the userdata
with a pointer to themselves as the first item on the stack. This
breaks the normal lua stack discipline for callback functions.

This commit makes it work normally, removing the stack entry for
the function before it is called, and also making a local copy
of the function pointer in case the lua garbage collector should
deallocate and destroy it when it is removed from the stack.


Former-commit-id: 1e68893443b87a8361414cd347512b0bdaa7c033
Former-commit-id: 10f2a36f3845a184ec88fbb4bb8311a6ad551dd8
Former-commit-id: 0fd6bcc659dcc40880858aba445c866d4c38ef61
Former-commit-id: d4e4b182be84621065c2003f6a92c7a046aca218 [formerly 8afef925327a98cbfdf3b83b4b2c66e171df35ab]
Former-commit-id: 94312f1c19a240644c1679c6e846b869a1bae98c