refactor game_lua_kernel not to use resources:: for gamestate
This commit changes the majority of the wesnoth callback functions
for the game lua kernel to be private methods of the game lua
kernel. The game lua kernel is given direct access, via it's ctor,
to references of many of the things provided by resources. The
call back functions are modified to access these private reference
variables instead of resources, and they are added to the
scripting environment using the lua cpp function mechanism.

This greatly improves encapsulation and will eventually lead to
eliminating the resources pointers from the game lua kernel
entirely.


Former-commit-id: 37bd8e6b795736b2649bf3e20960b2cad3f50535
Former-commit-id: 3ccf5507258aa7b03e4a5188d70575db25fb8ad7
Former-commit-id: a9e1865723b2deae6a7c440c34774934b0998c6b
Former-commit-id: b300618aff7da848a057f11ef2700d83cd294f10 [formerly 113a27149f8720cd419d388d336a9b8900d2429c]
Former-commit-id: 528df6b0ac1dddf6352fbf36e645b5b98edf1732