check properly if a config attribute value is not a boolean
I don't know if this commit is related to or fixes any bugs, but
there are many bugs reported about fog / shroud / mp campaigns it
seems, and these lines always looked a bit squirrely to me. Since
boolean config values can be indicated by "yes", "no", "true" or
"false", testing for the presence of "yes" / "no" only seems like
it won't work in general, even if it works oftentimes. This commit
replaces it with a test for whether the to_bool function is
selecting the default boolean value.


Former-commit-id: ca212711861ad26c513a3f5b5489cbda52297d99
Former-commit-id: 99667aff90eafb3ed2f5c7021d0a46f1dad7f68a
Former-commit-id: c1946eef3237026f43fd93d2691d35228a34da1e
Former-commit-id: 514b4f9df2d2b62a8461e184deb0d077d84ed815 [formerly a71a9976bb160262f448f925b8806f03cb90f0f0]
Former-commit-id: 89770e2d5a5d78714cc3ae05fd6e6841657c977b