check properly if a config attribute value is not a boolean
I don't know if this commit is related to or fixes any bugs, but
there are many bugs reported about fog / shroud / mp campaigns it
seems, and these lines always looked a bit squirrely to me. Since
boolean config values can be indicated by "yes", "no", "true" or
"false", testing for the presence of "yes" / "no" only seems like
it won't work in general, even if it works oftentimes. This commit
replaces it with a test for whether the to_bool function is
selecting the default boolean value.


Former-commit-id: 6d2fe8528e432826c30323c6c7b3d23a31343415
Former-commit-id: 6af11de9d1fe4f717b30abe904f9443efb595990
Former-commit-id: 0b270200a1a513d5e6a158589227fabf6a1c2b59
Former-commit-id: 4b3774fa8914c328b86fdf41cd51724a656d6b66 [formerly 4d86145604617b3cd3e2353553768b493bbc324c]
Former-commit-id: b01ca39dccb55851cd5dd2b64c11471dc4331302