add plugins context, manager, create lua thread object
This collection of classes provides a framework whereby the engine
may provide scripting capabilities in any "context". A C++
function may declare a plugin context by providing a list of names
function objects, then may call "play_slice()" on this context to
transfer control to lua. A lua script may generate some arguments
and request to call some of these, then yield to the engine. C++
will run them until one of them signals an error, or that the
context is now invalid (using a boolean return value).

The intended semantics is that lua may only make requests
corresponding to the current context. As soon as a new context
is used (play_slice() is called again) the old context is invalid
from lua's point of view -- calling those functions is a lua
error but not undefined behavior. This should be the case even if
the old context has not been destroyed yet in C++, for instance
if a lua request resulted in a new context being created. There is
a system of locks in place to ensure that when old contexts are
invalidated, stale requests are discarded.

We also add the coroutine library to our lua base. This because
threads are run as coroutines, and writing them requires the use
of coroutine yield from the standard library.


Former-commit-id: f1281e2d66ee0d56603a712bda67869284a13605
Former-commit-id: 62c63d1dc887d08ec440f95445ffbbc8a0904f56
Former-commit-id: 7ec91802f4761e0c1c9523f4b661609ee62f632f
Former-commit-id: ad0c08afee114f56dbdfe216c2fea77af889f081 [formerly c316459ecb3d984839b58aad83fa116d1f502d3d]
Former-commit-id: 2259880398e99e728b17b896e2e23672a14aba3e