add wesnoth.show_lua_console, as a debugging aide
This allows some sort of "drop into debug" mode for lua if your
script detects an error.

We could consider to actually add an "option to attach debugger"
as an error handling mechanism, but it might be complicated
because of wesnoth exceptions which must go through lua and be
treated as lua errors, and we cannot handle two exceptions at once
obviously, so the lua interpreter is probably not safe to run that
way in general. Anyways this commit does not do that.


Former-commit-id: 830a09b2ee59394792b77f8986077ecec001b31e
Former-commit-id: 736aa5d6868e45b2a84608d4a5ca45b29529c9f5
Former-commit-id: 0242d2c11019547ead00f057e8db4616480ccf06
Former-commit-id: 163ab4585ac0f5f90b0a6f78f91877108c67fdc8 [formerly 352e1ad7cdf42cacb1ad040ec86023dee28d32ec]
Former-commit-id: c47f36d2e76a56e9f5002f87642f5dbc075c2459