fixups: don't construct string from NULL, avoid complicated API
In the error handling routines for compiling and running a script
at runtime, don't construct a string directly from lua_tostring,
because if it is null that is a runtime error. Instead check for
this scenario.

In intf_require, when getting the "wesnoth" global, luaW_getglobal
is unnecessary and lua_getglobal should be prefered.


Former-commit-id: 60077990f1ff295c993e4a401c22a6101a04c4f4
Former-commit-id: 90610b8d78585dd1636f00ab87da2bd6db552d5b
Former-commit-id: b01fee9b7851af86d3ee28c04fcd5ceffc49d0fe
Former-commit-id: 6f7de82ce8d9593bcf1baefcf9f9043cdb59c048 [formerly c3b264cf53831b4d62be4a0d56cd852663db7647]
Former-commit-id: 0760f2922db2f731c5213b17848eba76a57ee9ea