in in-game load dialog, check health and version before exit game
This is fixes a minor bug:
When loading a save from the in-game load dialog, as soon as the
user clicks "ok" to load a savefile, an exception would be thrown
which effectively quits the game, before determining the integrity
and version info of that file.

It's better not to exit a running game if the new game can't be
loaded, so in this commit check these things and abort before
exiting if appropriate.


Former-commit-id: ed731da63171f4cffef6606398e49ac470b2bc81
Former-commit-id: b0d09aa851e726c38c9e80432229d561934580ed
Former-commit-id: 61d122ed246df46c39e4c38a6e97b7ca277c3b97
Former-commit-id: 255181923c23904b631bb13fa140c76810c4ee5e [formerly 163621ddc88f92f3dd48288d67320b60fcdc1186]
Former-commit-id: 057207875122d87448ad97d2d73edd7c8a506954