make force_lock_settings default to true for [scenario]
In most cases the campaigns designer don't want the use to be able to
overwrite scenario properites like "turn limit" or "use fog".

It is easiert to do this here in the code that loads the config than in
the code that reads that value becasue there are multiple codes that
read it and it's not clear that all of then know whether teh scenario
was a [multiplayer] or a [scenario].

Partly implements http://gna.org/bugs/?23037 which suggests to remove
the ability to overwite those settings in capaigns completely.


Former-commit-id: 7bf71e06f120d4867b3892838e4b4c54c67a1d84
Former-commit-id: 2f591818ddb7943ef1a6073d89f97a40f1b2a8a8
Former-commit-id: debb24a8edc6957d0eabd4e38cfa734a7d38479e
Former-commit-id: 470b427f17dbed213f3e80c85b99a187293337f9 [formerly 0672f9a7d3b0f2dedcdf27d8f234946ac2ecff0d]
Former-commit-id: 2985dc49542db01843a78a22f07285883a728f56