make game_events::pump not a singleton
Instances of this are now owned by the manager.

It would be nice to refactor this to provide all of the front-
facing methods in the manager instead of in the pump, since the
pump itself is somewhat an implementation detail. But this is
nearly the "smallest" diff change that would remove the singleton
restriction.


Former-commit-id: dee22be3c1860eb7a4b116aaeb4ac6329f04576f
Former-commit-id: a309f7c2491e6a9e53cb512ffa2311f68d48d710
Former-commit-id: e48ca3d48002c3c59ec1b10641d25e4669fa1014
Former-commit-id: 9f43fb86e6a71b247350d9be74b53437eddd282f [formerly e6c5025266d114f52bca831af8cfadc0f63ea3a7]
Former-commit-id: 155ab52c63db7f7030883dab84bcd50710e807a4