add documentation, scons check for jpg support in sdl_image
It was reported on Dec 22 2014 in irc that the game is broken if
this is missing, and a search shows that bigmaps and many campaign
story screens use jpg images.

It makes more sense imo to document this as an official dependency
now, and if we decide to change everything to png's then we can go
back and disable the jpg check then.

Conflicts:
	SConstruct

Former-commit-id: 1eab9fab4086bd247baa21ed8914cc98ef2ceb03
Former-commit-id: 69820b49be16d025408559a4e7909ef3fa55181e
Former-commit-id: dd88619e264d26ac10dc094e2624c78d8300a998
Former-commit-id: fe81ea8bbe82d7d725212226e71ffc740af12a00 [formerly 8403ad3fc18a6db0616f6ea5d4f5d85e22f1630e]
Former-commit-id: 0ce5283c5e1d0dd40d7aba17dd7311bd82f7434f