add documentation, scons check for jpg support in sdl_image
It was reported on Dec 22 2014 in irc that the game is broken if
this is missing, and a search shows that bigmaps and many campaign
story screens use jpg images.

It makes more sense imo to document this as an official dependency
now, and if we decide to change everything to png's then we can go
back and disable the jpg check then.


Former-commit-id: 81bd552ca6cd8a0b03089276528ae970fef59d42
Former-commit-id: 76863c1e77e20eaa3111446963102626043a46a7
Former-commit-id: da282096ceef04a18d8c8bd02b1730e43f5364c3
Former-commit-id: 3ad40def3b8c2548f05dd19dd78ecd954db3c7e3 [formerly c6e50df3d558a67d814d48cb9876d71823787652]
Former-commit-id: e1d0b935447a9d81e36b66fb38dc53726cdf8121