gui2/tgame_cache_options: Handle negative values from filesystem::dir_size()
The Boost.filesystem version of this function yields a negative value of
-1 to signify a non-critical error. For some reason, utils::si_string()
formats this incorrectly and even yields an incorrect number, probably
due to some floating point math mess-up.

Kind of fixes bug #22261 by preventing a call to utils::si_string() with
a negative value, but that function really ought to be fixed separately
anyway.


Former-commit-id: 6539ed82431425b116bd996c42cfa13c96e31f8c
Former-commit-id: 881e96c8fd2203ac0e10a9c0cb1429af498239ea
Former-commit-id: 257eb898ee9442555840a2ce4b4c6f657d85a7b0
Former-commit-id: 0952f1fd3d83f6d99cef738c9f70b6aaa9429c27 [formerly d877a2fdcb149bf4e46cae7bad3a4ba468b0c081]
Former-commit-id: 7c9ec79b47a0d62f6102bf0b1479b64815bff548