Use delete_file() to delete reg files in config_cache::delete_cache_files()
The reason is that, although both implementations of delete_directory()
can delete a regular file just fine, the Boost.filesystem implementation
depends upon a noisier implementation of get_files_in_dir(), which sends
an error message to stderr if the target is not a directory:

  20141223 21:54:36 error filesystem: Failed to open directory /home/shadowm/.wesnoth-1.13/cache/cache-v1.13.0-dev (cd02236-Modified)-aec97a1c3d958e534a2c379e200f91293932195b.gz: Not a directory
  20141223 21:54:36 error filesystem: Failed to open directory /home/shadowm/.wesnoth-1.13/cache/cache-v1.13.0-dev (cd02236-Modified)-b1a7724a622c9a107d2e5b70701d75e60db6ae4d.checksum.gz: Not a directory
  20141223 21:54:36 error filesystem: Failed to open directory /home/shadowm/.wesnoth-1.13/cache/cache-v1.13.0-dev (cd02236-Modified)-b1a7724a622c9a107d2e5b70701d75e60db6ae4d.gz: Not a directory

I really don't want to overcomplicate my code to deal with files and
directories separately, so I'm just sticking a check that incurs in an
extra stat() call for every entry to choose whether to use
delete_directory() or delete_file(). This is all very infrequently-run
code anyway, and modern OSes are smart.


Former-commit-id: 7079968714dcf9702fa3b2d4cb72a3b1bb809d36
Former-commit-id: e37547bd52886c554470ab2938876ce8247a1d65
Former-commit-id: de7a9f918382b21bac921795761e571d20c3b812
Former-commit-id: 371475fc155dc68cf3e6098f5925a67e2b1ef29b [formerly 4f9a1c3f2687b2223739fa11653ea53d4ef73ef5]
Former-commit-id: 36f4c8c3663a826059b87bb435c919ddc5a5490d