move WML [event] impl to lua, fixup core wml vs lua loading order
This commit moves [event] to be implemented in lua/wml-tags.lua.
It turns out that because of some questionable ordering in
data/core/_main.cfg, none of the tags defined in
data/lua/wml-tags.lua are actually defined at the time that core
is read, instead they are defined right after this. This is broken,
the entire wml library should be defined before core is read. Thus
we reorder some directives in data/core/_main.cfg to ensure this.

This commit adds lua callbacks `wesnoth.add_event_handler`,
`wesnoth.remove_event_handler` as well.


Former-commit-id: 3885c6940223c9a32665124ffae2192d8ae4493f
Former-commit-id: 8bfcf296b64b0bbdb8d17a8c2a6f7982ff8f8b2f
Former-commit-id: f1aa1b5c1a0da7350337621f0b709b388ff1aa4e
Former-commit-id: 81b27964f011da92df1b054995db7459e21822e4 [formerly 06fc5f65de39753d3a29164358e889fe9a99fd6a]
Former-commit-id: f6e4c572fd4a27b2b6967e22748499697e3c4076