lua: refactor to use lua 5.2 syntax for custom metatables
In lua 5.0, the preferred way to define custom metatables is to
use registry keys defined by pushing pointers as lightuserdata to
lua. (These pointers can be pointers to arbitrary constant objects,
as long as they are unique.)

In lua 5.2, the preferred way is to use constant strings as the
registry indices, and several lua API functions are introduced:
"luaL_newmetatable, luaL_setmetatable, luaL_checkudata,
luaL_testudata" which greatly simplify the operations for creating
the metatable, assigning it to a userdata, and checking if the
userdata has that type.

This commit switches over the keys associated to textdomains,
translatable strings, vconfigs, and lua gui2 dialogs to use the
new style, and simplifies some of the implementation details in
the process. For instance, luaW_toconfig no longer has to have an
optional argument to an index on the stack where the tstring
metatable is... these details are handled by the lua API instead.


Former-commit-id: 307873a1ee57d740247ee2d4d674a3afeeb80254
Former-commit-id: b74a10da99004b916f1a944cb040161091288a90
Former-commit-id: dffe4035c4c3755095b9fc1f97ac31c1d12317cd
Former-commit-id: 716df933331c87dfef10388eeae70733ca856d13 [formerly 7d01f2be2447be900f200e25b3ba4d0078cf69a3]
Former-commit-id: 80eb520ca74601665ba20c0ff7c4e96cc7f2288c