ilua: eliminate the "_" holds last executed command feature
the ilua `_pretty_print` feature has the property that it sets _
to be what was printed, as an interpreter feature so that _
represents the results of the previous expression. However it
appears that this may conflict with some parts of our api, for
instance the _ is used to hold the return value of
wesnoth.textdomain in the lua implementation of [harm_unit].

Note that there is no actual bug that I have observed, I merely
anticipate that this _ feature won't be compatible with our
established gettext conventions.


Former-commit-id: 8e111f3b35163c2a807ad7b95fa7d877fb7b61dc
Former-commit-id: e807b01924149f7a3301246c6b25082733b2d95a
Former-commit-id: c7ce4c4033b7316c58f110da14955756f1671828
Former-commit-id: 96dee560f8b711afb1a3c4cfc38f829be9020e4f [formerly 7c1c96c1f791e83212cdf9750a70cd1beec4bc75]
Former-commit-id: e65194bc6770d0ac32cb33b6e2bb29f5b5e63905