fixup 92461d04cc160542a31dfff1e3b37a459b87855b [formerly 80eb520ca74601665ba20c0ff7c4e96cc7f2288c] [formerly 716df933331c87dfef10388eeae70733ca856d13 [formerly 7d01f2be2447be900f200e25b3ba4d0078cf69a3]] [formerly dffe4035c4c3755095b9fc1f97ac31c1d12317cd] [formerly b74a10da99004b916f1a944cb040161091288a90] [formerly 307873a1ee57d740247ee2d4d674a3afeeb80254]
That commit overlooked the tstring_concat_aux function in
lua_common.cpp, a helper function for tstring concatenation. This
caused lua errors when HttT was started, arising in the code to
display objectives.

This commit fixes up the tstring_concat_aux function to more
directly check the tstring type.

It also fixes luaW_totstring and luaW_toconfig so that they don't
immediately flag a type error if they fail to convert userdata to
a tstring... this was caused by inappropriately using
luaL_checkudata when luaL_testudata should be used.


Former-commit-id: fad92c04928c2dd0bf623d4801a1244d7085aa57
Former-commit-id: 5dc8803885363528a38922e2589d36cb58178f4a
Former-commit-id: d61a54e09ccb0da96ca0ad87e04e4f78d3cffa70
Former-commit-id: a35818cb31b87f6f79c395ec9c4b8d5fa2cf0a2d [formerly c5bb56a0e033913aae6b8865b84786d1f7239db9]
Former-commit-id: 867d591e6e7ab503323c0eb13e0927cd5c3c46ab