cmake: Fix pot-update target breakage when bash isn't the default shell
(Default shell as in the /bin/sh executable.)

On Debian and other systems using a different Bourne-compatible shell
that doesn't support some of Bash's syntax extensions (in particular,
'&>' as a redirection operator targeting stdout+stderr), the pot-update
target is completely broken, stuffing the POTFILES.in lists for either
the last (wesnoth-utbs) or first (wesnoth-ai) textdomains with paths for
every C/C++ source file from our tree.

Shells that do not interpret &> the way Bash does consider it to be the
stdout redirection operator following an asynchronous invocation. Of
course such a thing is bound to cause weird issues like this.


Former-commit-id: b230f9bc6be8c59bf8253ae77aec102bb4b3e03e
Former-commit-id: e43a4a7487cc261a93c96164a2e163fdd4406421
Former-commit-id: 2cf4b95f4cdaaf1aab74bde48006b0131247b922
Former-commit-id: 5e9d00ddb7e3b5c68c3cfde157113804c1c36120 [formerly 3c7ec3ce896b76e9f296804bf2c4c12d05df35ed]
Former-commit-id: 7650d09f51afc5f2ee4729d6a56e50cde0d004e2