cmake: Fix pot-update target breakage when bash isn't the default shell
(Default shell as in the /bin/sh executable.)

On Debian and other systems using a different Bourne-compatible shell
that doesn't support some of Bash's syntax extensions (in particular,
'&>' as a redirection operator targeting stdout+stderr), the pot-update
target is completely broken, stuffing the POTFILES.in lists for either
the last (wesnoth-utbs) or first (wesnoth-ai) textdomains with paths for
every C/C++ source file from our tree.

Shells that do not interpret &> the way Bash does consider it to be the
stdout redirection operator following an asynchronous invocation. Of
course such a thing is bound to cause weird issues like this.


Former-commit-id: f9a9763211dfff7566394d4aa42bc018d7f5a00a
Former-commit-id: 2e4e4b44355a578ffa4286694c148a4d160861b0
Former-commit-id: c898ce935f8512a7205ac929cbdc6070b8da806d
Former-commit-id: 85e54b659076dce82fddb271d4e53d63040ea1ca [formerly 1e3bc1f0c91573234447e7dbc7cbf1169dfb87d1]
Former-commit-id: 43259e884a04d25a66090e34a5090c7771a18b4d