fully fix bug #23076
This commit keeps the various UI elements of the recall, unit list
dialogs in sync by passing a smart pointer to the underlying set
of units between them. This seems to fix all of the problems
mentioned in 23076, preventing assertion failures and mismatches
between the unit list and the preview pane.


Former-commit-id: 0b9906f91d5e3661004c2916a184a5d8021607a2
Former-commit-id: 5ef9694a3d622760d03f8d9626e8be6b1fb46b7d
Former-commit-id: 1ca34a73885f30af27da7adf053aa67d86597e18
Former-commit-id: 16879b7e9e3243857205353ff893ee4f82dd2f8c [formerly f30134acb32599fae07e70b3c893171f8b8b1129]
Former-commit-id: 54aa588c6cf72a801e0dc37775fa67023aaec797