Lua: Fix uses of `tostring` calls as LHSs to `or`
This commit changes occurrences of the pattern `tostring(x) or y` to
`tostring(x or y)` in the following Lua scripts:

  - `data/lua/wml-tags.lua`
  - `data/lua/wml/items.lua`

`tostring(x) or y` is unlikely to do what the author intended (and is
pointless unless `x` is an object of a custom type with unusual
behavior), because `tostring` returns a string (`"false"` or `"nil"`)
if `x` is a falsy value (`false` or `nil`), and all strings are truthy
in Lua — thus, `tostring(x) or y` will (unless `x` is of a custom
type) *always* evaluate to `tostring(x)`, never to `y`.

`tostring(x or y)` should, I expect, give the intended behavior.


Former-commit-id: 46bf4d7e1937be3ac0b884eea1a0ce18cb673b23
Former-commit-id: 022edf4d764a00c2cc4e49e6d4859dd4b381014f
Former-commit-id: 82e47b5af79bc936c9e67fbd9ea72a1764549b4e
Former-commit-id: 699bd3afdc4eafd9b6fef32ebec182e8c8d8a9d0 [formerly dd4572e1e468710c77522bc256086a29c72894b6]
Former-commit-id: 97588beca37fe9486df44990dc06f7a9b220cfb2