T S1: reduce chance of quintain for dying before turn 4
The reason for this is that if they player is lucky, they might kill him before turn 4,
and when it dies, Delf leaves, preventing the entire attack exposition from happening.
This cuts off the student's moves after they move to keep on turn 3, and adds a dialog line
by Delf regarding it.

This doesn't fix any situation where the player disregards instructions and doesn't return to
the keep, but that's their problem.


Former-commit-id: c7922e8337065043a246918f35c8a05fcd7c4c8f
Former-commit-id: dfda598d5a9f682e3d1878f2305177a8b9730d1a
Former-commit-id: 5accf9f31c3541198168302f76c2084433aa8b77
Former-commit-id: d4afefbfc67725b1a2d4d44a0c42c716f6a0d7ec [formerly 5d90b27495d7ec21c400dad8041c25b4bc420a59]
Former-commit-id: e8c86e9030f96476df2a8df4017f3a18cc5d06c0