lua: fixup d822151f9203219964b844fe9bb8b36bcd2f93b7 [formerly 528df6b0ac1dddf6352fbf36e645b5b98edf1732] [formerly b300618aff7da848a057f11ef2700d83cd294f10 [formerly 113a27149f8720cd419d388d336a9b8900d2429c]] [formerly a9e1865723b2deae6a7c440c34774934b0998c6b] [formerly 3ccf5507258aa7b03e4a5188d70575db25fb8ad7] [formerly 37bd8e6b795736b2649bf3e20960b2cad3f50535]
This commit broke the lua unit modify method for unit.side, the
reason was that I wasn't sure how to handle a safety check that
required a reference to the gamestate -- I'm not totally convinced
that units should not be allowed to be made in lua if there is not
an active gamestate.

However instead of commenting out this line, I should have
replaced it with a version that does the same thing without the
safety check... not sure why I didn't do this the first time.

This fixes lua errors that occur in the test scenario startup.


Former-commit-id: 25cac739208d27f1de894fd46b587f0f5cca4bb2
Former-commit-id: 1d567a58af053f98e6abca3738cd4408507c1e74
Former-commit-id: 5899641f3aa45161ea29ff61abe70f9df8ad2dd4
Former-commit-id: f385ff938b8aad61d354f39b8c3cf951cba830dd [formerly 2087827cb9753bdea35a0a564b2895e8ccbd6875]
Former-commit-id: 62dcbd7e3f3567c0f514f83f3fb7f58db2022765