don't do variable subsitution when parsing [team][village]
Becasue are no posibilities to change the variables or the positions of
the villages before this code happens, i cannot see any common usecase
for this feature, Also it's not documented in the wiki.

Note that it is of course still possible to change village owners
depending on variables in a prestart event.


Former-commit-id: 05a6b6e183052d7389ae956263b3795f28b1dad9
Former-commit-id: 3aa4f46346b49032daa0c633bcf3e016c11b3b55
Former-commit-id: 2d0ab6c5e5d9e90300977324eb00070a2ca177fb
Former-commit-id: a6adf08e8b951845859221bf4d4e4698f7e82730 [formerly e88c2175410e574a2277402d32a88c91c6c8ee21]
Former-commit-id: 80417bd45b972ca0f6cc9be23994b930dbb7258f