cosmetic: tidy up punctuation in controller ctor initializer lists
Moves the : and , to lead the lines rather than end them. This
makes the initializer lists a bit more readable, as it lines up
these punctuations and separates them from the initializer
expressions. It seems to be mordante's style in most of the GUI2
code to do it this way. It also has the advantage that each line
is paired with its matching punctuation mark, so if you need to
add or rearrange things, you never need to edit a line solely to
move a comma around, reducing the noise in the diffs.


Former-commit-id: bad9aae5e61fe3b32c4b21858f3bdae6d7e84694
Former-commit-id: 5cb92cd45caa539fae653893c9d8635b2ba3edd6
Former-commit-id: cad4fc3c8fb47673e2f78a0b400e15305a41e698
Former-commit-id: beaa760b2a03da4325acdd14ea256b8f6637bc1f [formerly 7405374e7b47b736474e5cf14ef1759030d6d702]
Former-commit-id: a6db63c1e08ab6b5aa8bf45be22016d4e00ac705