lua: change a reference variable to a non-reference type
This line was not an error -- it is legal in C++ to bind a const
reference to a temporary value. It's only slightly more confusing
than usual in this case because the string it is built from is
a C-string owned by lua, and the temporary std::string object is
implicit.

I don't believe this commit will even make any difference in the
compiled code, it's only slightly simpler and more readable.


Former-commit-id: c5ec42c700fdf9d3e5480f6fdfea038af2d10e3a
Former-commit-id: 5b57665dcfba91934b471be821c6b459b6ffce5a
Former-commit-id: 4c1ad7245b972126f63e0e1d6dfad00be3badf5b
Former-commit-id: 72f735644bf465e8e73b5f865befa529fb573b8a [formerly deaa42e7ba2e63b3e3f89ded41c90950ecf2b905]
Former-commit-id: 0be5596fafa85465c7a0123d201d11484ab5fd26