add oos debug mode and mp_debug_checkup class
The game sometimes does some checkup to test whether the calculated results in a replay match the ones calculated during the original game.
This data was stored in the replay inside the [command] for that action. The problem is that this doesn't work in networked mp because we often send the [command] before calculating the results.
I added an alternative mode that used get_user_choice to compare the results, this also works in networked mp but it causes a little more network traffic.


Former-commit-id: 5ebdb5e500b25800329c1be9636a4ea1c5291fed
Former-commit-id: c02ab7774c3bd7eab602dedc4bcaa10c6e54f17b
Former-commit-id: 86c83e73cade3f847ab256ef8211b86657d9b9ff
Former-commit-id: 83fe095fd9ecab98945b0b02da88c2e9bf693b86 [formerly 1bcf4a95aa89860c6d007c61cfbb0fdcc4b70c65]
Former-commit-id: 99ea21704b7e7efab246a2cb2dc23cbd0c4523c2