preproc: Use the 'preprocessor' log domain for all diagnostics
We formerly used 'config', which is also used by the parser. This
resulted in a massive amount of noise being generated when trying to
obtain more details about the preprocessor's actions, since 'config' is
the go-to domain for pretty much everything involving WML (not even just
the parser). Since the introduction of --preprocess added a dedicated
log domain used by functions related to that switch, we might as well
stick to that.


Former-commit-id: 32a0da9a5c1b81d9ba0dfc7b9e7d6229afc3bbfb
Former-commit-id: b6a0f796c82c6f0a127119596052545769195743
Former-commit-id: 5782c4ab98715715fe02a3a6bcf427778328cced
Former-commit-id: 61119a82a25a3ae3f2bdea3d9148615464f74135 [formerly 7cf907af5153931c81db9fd989d95af173edd283]
Former-commit-id: 63ee240d574213846e7d0899c3a3db59d98e4d91