campaignd: Catch utf8::invalid_utf8_exception within handle_upload()
Otherwise add-ons with invalid UTF-8 sequences in the add-on name or the
fields matched against the blacklist (description, author, email, title)
crash campaignd since the exception is not handled anywhere.

We specifically want to handle this exception deep within
handle_upload() and with more specific knowledge of the circumstances
because otherwise we can't give the user an error message with enough
information to help them.


Former-commit-id: 14b5761a9970cd37f97901fe8d13bedf7662645b
Former-commit-id: 67b65c5ed61511eb68399c956040e195cb91c0f5
Former-commit-id: d2c7a04f9983eee41edb1efb470d0acca14ddd1c
Former-commit-id: 770c698e6fc8b9115c7d68e455db39d6e6b78724 [formerly 93bf2fd4ea8133eec94b2d5951c808be51d62e2f]
Former-commit-id: 10eca447785354050cb75b5f5d932caf8884d681