campaignd: Catch utils::invalid_utf8_exception within handle_upload()
Otherwise add-ons with invalid UTF-8 sequences in the add-on name or the
fields matched against the blacklist (description, author, email, title)
crash campaignd since the exception is not handled anywhere.

We specifically want to handle this exception deep within
handle_upload() and with more specific knowledge of the circumstances
because otherwise we can't give the user an error message with enough
information to help them.


Former-commit-id: 3eb3291e9664c7f6d6d7258df98818d0363fa40a
Former-commit-id: 841e84b3e5bd5e827491201e567e98d4a475027d
Former-commit-id: c739080b59d7cf617aa59148509d8763f880bac4
Former-commit-id: 064e6481f0edf0ebe3840f01987d1a80e64852e5 [formerly 8a4d1b633fa96d3af660837b79b90b28eed68465]
Former-commit-id: 01a2177543711b001e734161646c8a6b4e863ce4