don't reload config when selecing campaigns
this reverts
https://github.com/wesnoth/wesnoth/commit/8b0b5fc74a94dcfcaa8479bdf1547008cae140af [formerly 4463040209f0bdac9e39c4fdf7b637190a54a026] [formerly 51cd072f65287e1c128889a61565bf94b2db2ece [formerly 9f602f9e2abe74ff4859f1cd83f8b6e44946b035]] [formerly 9fc8b0e5dc9fe29ce4082cd1a15e2f141003ebe8] [formerly e7e06056b9ffb440d2f20fd7bd0842eda6f7a3da] [formerly 8503e7611e4225635e4f89699ff9d259b52c1a46]
and fixes http://gna.org/bugs/index.php?23229 and
http://gna.org/bugs/?23038

This means that [era]s and [modification]s inside an '#ifdef
MULTIPLAYER' might be loaded during the sp campaign selection. But
that's the smaller pain compared to the wml reload. A proper fix for
that one would be type= attribute for [era] and [modification] just like
for [campaign].


Former-commit-id: 22af1dd0ac9b68812307b316084447c86eed0f7c
Former-commit-id: d1d61ddd93c184877acc3eebb07e018be52a78dc
Former-commit-id: ec8218c5eba52faa2f721bba37bd63535629ff5e
Former-commit-id: 239f1920ad3bbcc13376cd92f818377abe1a3dac [formerly 7cc8a47b4fa18d7714d6a8eb6830c57f94a0acc7]
Former-commit-id: 3719c4630919ea2e24e9ee6627c4e48ecc7f6690