add side specific results in [endlevel]
adds [result] tags in [endlevel] as suggested in
http://gna.org/bugs/?4960
the following [endlevel] attributes can now be given side spcific:
'result', 'bonus', 'carryover_precentage', 'carryover_add'

In the normnal case we show a victory message on a client if any of the
local humans is victorious, but we have to watch out for border cases
like when there are no human sides and behave just like we did
previously in this case.

This also fixes a bug that the calculated carryover could be different
from what is shown in the message (see comment in playcampaign.cpp that
was removed by this commit)

This also removes some unused functions in carryover.cpp that were
incompatible with this commit.


Former-commit-id: 8c6969fa244826b3dd92fe031269d950e85af61d
Former-commit-id: c5f514fd9bdebfb26d181b5541b06ca673e7d1a9
Former-commit-id: fe15f14ebc635d8c21031631e732bdf9f51f2828
Former-commit-id: 05c416e7cdd848f5683b938181a55d25599a6b7c [formerly 74dbb51daa1fa3ca0f27429dabac3f37c920b0f5]
Former-commit-id: 70cab4e98639a51fa587ec01a2a6396212bcd341