cleanup cfg_ member of unit class.
the cfg_ field of the unit class now doesn’t store the advancements and
the abilities anymore. Instead they are stored in separate fields.

I used boost::ptr_vector<config> because that’s similar to how config
objects store their child values internally. Unfortunately
boost::ptr_vector defines ptr_vector::value_type in a bad way so that it
cannot be used for BOOST_FOREACH loops value types, so i had to
change lua's push_check file to use T::reference to be compatible
with ptr_vector

This commit removes all child tags of the unit cfg_ because the only
tags that are now left in cfg_ are [variation] and [advancefrom] from
[unit_type], which werent used in [unit] and only caused the savefiles
to bloat.


Former-commit-id: 9012c0bb0676c040f8cadff19fe054473696d59c
Former-commit-id: 82dd576d4166ed105a34a4c3eb0e3da79e698cad
Former-commit-id: e7f8661c8753f4267c99587b82de55e9ab92a8b2
Former-commit-id: 2da6a46377d8e2ea5586ecfd784dff16d98b62b0 [formerly 928661809a6105e0e7cb9609e5698dfbcda9c142]
Former-commit-id: 4728546ae17414ffdc87c64ed4dbef3d621ae4a5