cleanup cfg_ member of unit class.
the cfg_ field of the unit class now doesn’t store the advancements and
the abilities anymore. Instead they are stored in separate fields.

I used boost::ptr_vector<config> because that’s similar to how config
objects store their child values internally. Unfortunately
boost::ptr_vector defines ptr_vector::value_type in a bad way so that it
cannot be used for BOOST_FOREACH loops value types, so i had to
change lua's push_check file to use T::reference to be compatible
with ptr_vector

This commit removes all child tags of the unit cfg_ because the only
tags that are now left in cfg_ are [variation] and [advancefrom] from
[unit_type], which werent used in [unit] and only caused the savefiles
to bloat.


Former-commit-id: f731754a6b0c2161f21435fbd1392ca4228c3ca4
Former-commit-id: a52bea44e27caa086bbca12a940cde52b6403856
Former-commit-id: 7b8eeb1e201b796db09f971d1a703dc5952ed595
Former-commit-id: fa37c6cec262cf558d21ce3e646e170a311d2337 [formerly 0e0e3b1fd74dce3787ee0b99a38caff9cc2e152e]
Former-commit-id: d0c6725aaff59fcab172a2d6f1e64dca8496eb02