don't do auto delete save on defeat in mp campaigns
We now use 'end_level.proceed_to_next_level' instead of 'res == VICTORY
|| (game_type == game_classification::MULTIPLAYER && res == DEFEAT'  to
decide whether we do auto save deletion. The intention is to adress 'if
MP campaigns ever work again, we might need to change this test'
as it was commented in the code

We also refactor playsingle/mp_scenario to not do
playcontroller.update_savegame_snapshot(); if res == DEFEAT since it is
not neeeded in this case.


Former-commit-id: 9c9f94df687441067471b86a6719c7e8726e0808
Former-commit-id: 325948d43322b9e233fe7d6d2656fcafd34a50cd
Former-commit-id: 07861b430aad2ee496409da714f1444c548997a2
Former-commit-id: 34370fc6a1a00ecfb5300b7577ffa9492b78a80a [formerly f093a61a8bf669bd83d4662ad6961b18d7c77720]
Former-commit-id: 0336cb62b90ce3742b9df232af7c28f2b97b8922