fix zoom slider focus issue (add a separate class for zoom_slider)
When the slider widget was added to ThemeWML, some work was not
done apparently to make it play nicely with regards to focus and
key events with the other theme wml widgets, causing it to retain
focus for keyboard events indefinitely. At the time this was
resolved by disabling these events, but this cripples the many
other sliders in the game. In this commit we introduce a subclass
zoom_slider which has this functionality disabled, and use that
for zoom sliders, so that we can get the desired functionality
back for the other sliders. See parent of this commit for more
info.


Former-commit-id: 6af57f5ca0c6ca9970fdaee6757a148c19adb3dd
Former-commit-id: 536812abde75a64672ae907ee6ba6d2463bdf55d
Former-commit-id: 45be3e6080f3eeccc92b6ff81475442f205fe3a3
Former-commit-id: f1adc6ac1b913b0df56e09e6b65c3add8b099af0 [formerly b0ebaf2f8e6f49cfb1130daed13149c9602f46f0]
Former-commit-id: ea8a932116ebb47c35aade830aec192f0e28ca21