fix zoom slider focus issue (add a separate class for zoom_slider)
When the slider widget was added to ThemeWML, some work was not
done apparently to make it play nicely with regards to focus and
key events with the other theme wml widgets, causing it to retain
focus for keyboard events indefinitely. At the time this was
resolved by disabling these events, but this cripples the many
other sliders in the game. In this commit we introduce a subclass
zoom_slider which has this functionality disabled, and use that
for zoom sliders, so that we can get the desired functionality
back for the other sliders. See parent of this commit for more
info.


Former-commit-id: fe1352d79d703d63029abcab8fb764a7538bf54f
Former-commit-id: e3a1d15b6459228563a8d3f09a7bbf900a57b87d
Former-commit-id: bbb938b57f1e30ef141905a6e887d65ed7749de0
Former-commit-id: 081d248229cda3a25faebc336d9d09a609553a17 [formerly 5a463785ccb2d0c13f4986f83eaf607613c76e59]
Former-commit-id: 9ed9cd40bdfd66c6bb0bcf82a2fdb109da254c2c