recorder writes directly into saved_game object
Previously recorder had config memaber and saved_game had a config
memeber.
And when saving a game the config was copied from recorder to saved_game
and the other way when loading a game.

Now the recorder object directly writes into the saved_game object. This
saves some copying when saving and loading data.

I also moved the pos_ variable from the recorder object to the
saved_game replay_recorder_base object, This fixes a bug where saving a
game during a replay also caused to not yet played turns to be written
into the savefile.


Former-commit-id: a9c6d7a9f713dacbf66794b71c73692b8d2e4808
Former-commit-id: 74c819103c743c31200e8f93237fabbde581c185
Former-commit-id: 6c73fc1d88bc017324e6bb1dbc715e5afc0ee233
Former-commit-id: e4681b6435b7cbcd9de85bb6bdccc7280a540b72 [formerly 9c59a0261b1b33923f7f682d14ef9cce80d7074b]
Former-commit-id: c21e6d7b9cc7090b08872ca72a50d9c73f7d2c30