fixup server controller tweaks, dont tweak host
In 1.11.x cycle, we moved all "client-side controller tweaks" to
be effected by the server, so that clients won't on their own
change the data they find in a level config at game start. This
had some benefits:
- Only host and server are ever modifying the level data
- Easier to keep observers and clients synced correctly
- Easier to change client behavior with server side patches

The server-sided tweaks happened at the following point in time:
1. Host clicks start game button, a [start] signal is sent to
server.
2. Server now checks the level config and tries to make sense of
it, making sure every side has an owner and such. Then it issues
controller change commands to all clients, making sure they match
its picture of what is going on.
3. Server passes the [start] signal on to the other clients.

This introduced some bad behaviors, however, because the host
does not wait to recieve [start] and instead goes right into the
game.

- When host gets controller change signals for AIs it controls, it
would overwrite the names of their leaders with the host names.
(bug #23021)
- It would cause some GUI flashing at game start for the host
(bug #21156)

In this commit, we make sure not to send controller tweak messages
to the host about sides that it controls.

We still may send messages to the host about sides it doesn't
control. If there is the possibility that a game may start and
an AI side is assigned to someone other than the host, then this
commit won't fix the resulting renaming problems. (However afaik
that cannot happen at this revision.)

Conflicts:
	src/server/game.cpp

Former-commit-id: 5eb7648066a41580df8cb4a7368d7aeee0d986ad
Former-commit-id: 2d52a7f272d1511bbe2cb0daba03a94c2778866b
Former-commit-id: 45b6ccf2bf9ee19df9fec1b8c06bcd2b758996fc
Former-commit-id: d1fd156f0e1ee3bd9bc4cf8e9f514eee9977c8ed [formerly f8d6a50fb9da18f59439e42c85f697183f228602]
Former-commit-id: 4019ada34dd3db078cdd5007636f5b8a7fdc15de