add support for styled fonts outside of SDL TTF in GUI1, help
This commit allows bold and italic font files to be specified in
the fonts.cfg file. When attribute "bold_name" is set to the name
of a font file, this font will be used in place of SDL TTF bold-
styling applied to the base font. Similarly for "italic_name".

This commit assumes that the code points for the styled versions
are the same as for the unstyled version. If that assumption is
not true for some font that we care about then this commit will
likely have to be revised.


Former-commit-id: e56aba2d44af89f61a96270d15bcfb0424b18027
Former-commit-id: d9ce00730d6e27fd6d39252ed4667cc270b9f005
Former-commit-id: d589e4afd131178531993d9c5894218cd3d60cf4
Former-commit-id: bcc43436ddf626e5f638799c398d09790801a2b0 [formerly c5f9fd9bd486953e73b332c655819c924ccfdd09]
Former-commit-id: 8bcf6115ec1f3797675e6ae810fe7a525a92b23f